% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astext.R
\name{astext}
\alias{astext}
\title{As Text}
\usage{
astext(x)
}
\arguments{
\item{x}{Vector to represent as text.}
}
\value{
Vector represented as a character.
}
\description{
Prints a vector as text you can copy and paste back into the code. Helpful for copying vectors into code for testing and validation.
Author: Bryce Chamberlain.
}
\examples{
astext( c( 1, 2, 4 ) )
astext( c( 'a', 'b', 'c' ) )
}
