% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{begin}
\alias{begin}
\title{Begin}
\usage{
begin(
  wd = NULL,
  load = c("magrittr", "dplyr"),
  keep = NULL,
  scipen = FALSE,
  verbose = TRUE,
  repos = "http://cran.us.r-project.org",
  runpath = NULL
)
}
\arguments{
\item{wd}{Path to set as working directory. If blank, the location of the current file open in RStudio will be used if available. If FALSE, the working directory will not be changed.}

\item{load}{Packages to load. If not available, they'll be installed.}

\item{keep}{Environment objects to keep. If blank, all objects will be removed from the environment.}

\item{scipen}{Do scientific notation in output?}

\item{verbose}{Print information about what the function is doing?}

\item{repos}{choose the URL to install from.}

\item{runpath}{folder or file specified}
}
\description{
Perform common operations before running a script. Includes clearing environment objects, disabling scientific notation, loading common packages, running fun/ or functions/ folders, and setting the working directory to the location of the current file.
}
\examples{

begin()
}
