% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.capture_warning}
\alias{cache.capture_warning}
\title{Capture Warning}
\usage{
cache.capture_warning(w)
}
\arguments{
\item{w}{Captured warning passed by withCallingHandlers.}
}
\description{
Utility function for capturing warnings.
}
\examples{
# this will only have an effect if a current cache exists.
\dontrun{
  if(!cache.ok(1)) withCallingHandlers({
     x = mtcars # base-R dataset.
     x = mtcars # base-R dataset.
       warning('warning 2-1') # this is the first warning we need tdo capture. 
       warning('warning 2-2') # this is the first warning we need tdo capture. 
       save.cache(x) # we'll capture it inside svae.caceh
  }, warning = cache.capture_warning)
}

}
