% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Concatenate.}
\usage{
cc(..., sep = "")
}
\arguments{
\item{...}{Arguments to be passed to paste0. Typcially a list of vectors or values to be concatenated.}

\item{sep}{(Optional) Separator between concatenated items.}
}
\value{
Vector of pasted/concatenated values.
}
\description{
Shorthand function for paste.
Author: Bryce Chamberlain.
}
\examples{
cc( 1, 2, 4 )
x = data.frame( c1 = c( 1, 2, 4 ), c2 = c( 3, 5, 7 ) )
cc( x$c1, x$c2 )
cc( x$c1, x$c2, sep = '-' )
}
