% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear.cache}
\alias{clear.cache}
\title{Clear Cache}
\usage{
clear.cache(cache = NULL)
}
\arguments{
\item{cache}{The cache list to clear.}
}
\value{
FALSE if a cache info list item is passed in order to assist other functions in returning this value, otherwise NULL.
}
\description{
Clears all caches or the cache related to the passed cache info list.
}
\examples{
# this will only have an effect if a current cache exists.
clear.cache()

}
