% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_cols.R
\name{comp_cols}
\alias{comp_cols}
\title{Compare similarly named columns}
\usage{
comp_cols(x, id_cols = NULL, pattern = "^i\\\\.", col_list = NULL)
}
\arguments{
\item{x}{a data.frame.}

\item{id_cols}{any ID columns to retain in each data.frame output.}

\item{pattern}{a regular expression pattern to match the distinguishing prefix/suffix}

\item{col_list}{a list of grouped column names, which takes precedence over \code{pattern}.}
}
\value{
A list of data.frame objects.
}
\description{
Extract columns with similar names and output a list of separate data.frames for easy comparison.
The presence/absence of a prefix/suffix identifies similar colums,
and the default corresponds to default-data.table joining (e.g., \code{i.count}, \code{count}).
You can provide a list of column names for extraction instead.
}
\examples{
x <-  c("5", "12", "zero", NA, "0")
check <- detect_na_intro(x, as.numeric)
data.frame(x, check)
}
