% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_dir.R
\name{cur_dir}
\alias{cur_dir}
\title{Find the current directory in RStudio}
\usage{
cur_dir()
}
\value{
script file directory path
}
\description{
Returns the current directory of the file, if the IDE is RStudio.
Cannot be used if the file is sourced. Use cur_dir_source() instead.
}
\examples{
# Add this in your script to change current directory
setwd(cur_dir())
}
\seealso{
\code{\link{cur_dir_source}}, \code{\link{setwd_cur}}
}
\author{
Kelli-Jean Chun, \email{kjchunz@gmail.com}
}
