% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_dir.R
\name{cur_dir_source}
\alias{cur_dir_source}
\title{Find the current directory of a sourced file}
\usage{
cur_dir_source()
}
\value{
sourced script file directory path
}
\description{
Returns the current directory of the file.
Can only be used if the file is sourced.
Use cur_dir() if you are not sourcing and working in RStudio.
}
\examples{
# Use when sourcing your script to change current directory
setwd(cur_dir_source())
}
\seealso{
\code{\link{cur_dir}}, \code{\link{setwd_cur}}
}
\author{
Kelli-Jean Chun, \email{kjchunz@gmail.com}
}
