% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duped.R
\name{detect_duped}
\alias{detect_duped}
\title{Test for duplicates}
\usage{
detect_duped(..., sep = "-^-", incomparables = FALSE)
}
\arguments{
\item{...}{vectors to concatenate.}

\item{sep}{a string used to when concatenating the vectors. See \code{\link[base]{interaction}}.}

\item{incomparables}{FALSE or a vector of incomparable---i.e., never-duplicate---values. See \code{\link[base]{duplicated}}.}
}
\value{
A logical vector corresponding to values that are duplicates \emph{or are duplicated}.
}
\description{
Test for duplicates in one or a combination of vectors.
}
\details{
This function \strong{\emph{should not}} be used in places of \code{duplicated()} and \strong{\emph{does not return duplicates}}.
Instead, \code{detect_duped()} executes the common pattern

\code{vector \%in\% vector[duplicated(vector)]}.

Providing multiple vector arguments to \code{...} concatenates the vectors element-wise
before testing for duplicates in the new vector.
}
\examples{
state <- c("CA", "CA", "FL", "CA", "FL", "FL")
cd    <- c(22, 11, 22, 22, NA, NA)

data.frame(state, cd,
           dup = detect_duped(state, cd),
           dup2 = detect_duped(state, cd, incomparables = NA))

}
\seealso{
\code{\link[base]{duplicated}} and \code{\link[base]{interaction}}.
}
