% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\title{Get Data Dictionary}
\usage{
dict(
  x,
  topn = 5,
  botn = 5,
  na.strings = easyr::nastrings,
  do.atype = TRUE,
  ecopy = FALSE
)
}
\arguments{
\item{x}{Data Frame or Data Table.}

\item{topn}{Number of top values to print.}

\item{botn}{Number of bottom values to print.}

\item{na.strings}{Strings to consider NA.}

\item{do.atype}{Auto-determine variable types. If your data already has types set, skip this to speed up the code.}

\item{ecopy}{Use ecopy function or agument to copy to the clipboard so that it can be pasted into Excel.}
}
\description{
Get information about a Data Frame or Data Table. Use getinfo to explore a single column instead.
If you like, use ecopy function or agument to copy to the clipboard so that it can be pasted into Excel. Otherwise it returns a data frame.
Author: Scott Sobel. Tech Review & Modifications: Bryce Chamberlain.
}
\examples{
dict(iris)
}
