% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_excl.R
\name{distinct_excl}
\alias{distinct_excl}
\title{Variation of the distinct() function to enumerate vars to exclude}
\usage{
distinct_excl(.data, excl_list, ...)
}
\arguments{
\item{excl_list}{character vector of names to be excluded from distinct calcs}

\item{data}{Data frame to be used (can be piped)}
}
\value{
Date frame returned with only distinct rows
(not counting vars that were excluded)
}
\description{
Variation of the distinct() function to enumerate vars to exclude
}
