% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drows.R
\name{drows}
\alias{drows}
\title{Get Rows with Duplicates}
\usage{
drows(x, c, na = FALSE)
}
\arguments{
\item{x}{Data frame.}

\item{c}{Column names as a character.}

\item{na}{Consider multiple NAs as duplicates?}
}
\value{
Rows from the data frame in which the column is duplicated.
}
\description{
Pulls all rows with duplicates in a column, not just the duplicate row.
Author: Bryce Chamberlain.
}
\examples{
ddt = bindf(cars, utils::head(cars, 10)) # create duplicated data.
drows(ddt, 'speed') # get the duplicated rows.
}
