% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_add_bins.R
\name{ezr.add_bins}
\alias{ezr.add_bins}
\title{Add Bins}
\usage{
ezr.add_bins(dataset, column, style = "equal", n_breaks = 10,
  fixed_breaks = NULL, new_col_prefix = "bucket_", round_breaks = 0)
}
\arguments{
\item{dataset}{dataset}

\item{column}{column.  It should be numerical}

\item{style}{'fixed','equal','quantile','pretty', or 'percentile'.}

\item{n_breaks}{number of breaks.  Only matters if not fixed}

\item{fixed_breaks}{the fixed breaks.  Only applicable if style is 'fixed'}

\item{new_col_prefix}{Default prefix is 'bucket_'}

\item{round_breaks}{number of digits to round too}
}
\description{
Bin a continous field into fewer values using variety of methods.
}
