% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_auc.R
\name{ezr.auc}
\alias{ezr.auc}
\title{Calculate AUC from FPR/TPR or from Truth & Prediction}
\usage{
ezr.auc(dataset, fpr_or_truth, tpr_or_pred, use_truth_and_preds = TRUE)
}
\arguments{
\item{dataset}{dataset.}

\item{fpr_or_truth}{either the FPR or the truth value.}

\item{tpr_or_pred}{either the TPR or the prediction value}

\item{use_truth_and_preds}{Default is TRUE - expects 0s and 1s for truth and a numerical value.  Set to false if you already have the TPR and FPR.}
}
\value{
returns the AUC.
}
\description{
Get the AUC.   You can input either a truth & a prediction or the TPR & FPR values.  Works on both H2o Dataframes and regular R dataframes
}
