% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_plot_auc.R
\name{ezr.auc_plot}
\alias{ezr.auc_plot}
\title{AUC Plot}
\usage{
ezr.auc_plot(dataset, truth, prediction, title = NULL)
}
\arguments{
\item{dataset}{dataset}

\item{truth}{The target.  Should be 0 and 1}

\item{prediction}{Prediction value.  Higher should = better}

\item{title}{Null by default}
}
\value{
Returns an ROC AUC plot
}
\description{
Generates an AUC plot:  TPR vs FPR.
}
\examples{
 ezr.auc_plot(dataset = easyr::dataset_telco_churn_from_kaggle, truth='Churn',prediction = 'TotalCharges')
}
