% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_audit.R
\name{ezr.audit}
\alias{ezr.audit}
\title{Data Audit}
\usage{
ezr.audit(dataset, custom_metric = NULL,
  export_name = "data_audit.csv", run_nzv = TRUE, nzv_ratio = 50)
}
\arguments{
\item{nzv_ratio}{}
}
\description{
Runs a data audit on a dataset.
}
\examples{
ezr.audit(dataset = iris)
}
