% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_binarize.R
\name{ezr.binaryize}
\alias{ezr.binaryize}
\title{Make Numerical Data 0 or 1}
\usage{
ezr.binaryize(dataset, column, equal_or_greaterthan = 0.5,
  factorize = FALSE)
}
\arguments{
\item{dataset}{Dataframe}

\item{column}{column in quotes}

\item{equal_or_greaterthan}{Converts this value or greater than to 1}

\item{factorize}{Default is FALSE.  Otherwise returns as numeric 0s and 1s}
}
\value{
Input dataset with modified column
}
\description{
Convert values greater than or equal to a value to 1 and all other values to 0.  Converts NAs to 0s.
}
