% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_capture_rate_dual.R
\name{ezr.capture_rate_dual}
\alias{ezr.capture_rate_dual}
\title{Capture Dual Values at a time}
\usage{
ezr.capture_rate_dual(dataset, target, score, n_perc = 0.1,
  target2 = NULL, higher_morelikely = TRUE,
  return_entire_table = FALSE)
}
\arguments{
\item{dataset}{Dataframe}

\item{target}{Target.  Should be 0 or 1.}

\item{score}{Score that sorts for the capture rate.}

\item{n_perc}{what percentile do you want to examine?}

\item{target2}{should be 0 or 1.  A secondary thing to measure.}

\item{higher_morelikely}{TRUE by default.  This is generally what you expect.}

\item{return_entire_table}{Return ntiles from 0-100.  Not yet implemented}
}
\description{
Get the capture rate depending upon the value of a score.  Requires that the target be 0 or 1.  It can be either numeric, character, or factor.  Target2 is a 2nd option where you can compare two metrics side by side.
}
\details{
Almost duplicates gains-lift table, but allows for side by side comparison.
}
