% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_collapse_binary.R
\name{ezr.collapse_binary}
\alias{ezr.collapse_binary}
\title{Collapse Binary Columns into Single Categorical Column}
\usage{
ezr.collapse_binary(dataset, newcol_name, column_names_to_collapse,
  make_factor = TRUE)
}
\arguments{
\item{dataset}{Dataframe}

\item{newcol_name}{Newcolumn name}

\item{column_names_to_collapse}{vector of column names to collapse.}

\item{make_factor}{Make factor or leave as character.  DEFAULT is TRUE.}
}
\value{
Original dataset with columns collapsed into factors
}
\description{
Collapse Binary Columns into Single Categorical Column
}
