% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_compare_vects.R
\name{ezr.compare_vects}
\alias{ezr.compare_vects}
\title{Compare values in a vector/dataframe
 Compare the column names or values in a vector. Determine which is present in one and not the other or present in both Allows comparing of passing df and a vector directly by handling the column name conversion to vector within function.}
\usage{
ezr.compare_vects(x, y, only_return_in_x_not_y = FALSE,
  only_return_in_y_not_x = FALSE, only_return_in_both = FALSE)
}
\arguments{
\item{x}{Either a vector or a dataframe}

\item{y}{Either a vector or a dataframe}

\item{only_return_in_x_not_y}{Default is FALSE along with other params.  If all are false, then
returns a list of each value.  Otherwise only 1 can be true.  Returns the vector of names that is in x but not in y}

\item{only_return_in_y_not_x}{See above}

\item{only_return_in_both}{See above.}
}
\value{
Returns either a vector or a list given params.
}
\description{
Compare values in a vector/dataframe
 Compare the column names or values in a vector. Determine which is present in one and not the other or present in both Allows comparing of passing df and a vector directly by handling the column name conversion to vector within function.
}
