% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_correlations.R
\name{ezr.correlation}
\alias{ezr.correlation}
\title{Correlation}
\usage{
ezr.correlation(dataset, target = NULL, only_variables = NULL,
  exclude_variables = NULL, only_numeric = TRUE,
  corr_method = "pearson", return_raw_numbers = TRUE,
  return_plot = TRUE, title = NULL, n_limit_using_target = 15)
}
\arguments{
\item{dataset}{dataframe}

\item{target}{optional.  The target y value you care about, providing cleaner, streamlined results}

\item{only_variables}{optional.  Only compute correlation with these variables}

\item{exclude_variables}{optional.  Do not compute correlation with these variables}

\item{only_numeric}{Default TRUE.  Only numeric values are considered.  Right now the rest is not yet implemented.}

\item{corr_method}{Default 'pearson'.  Other valid values are 'spearman' and 'kendall'}

\item{return_raw_numbers}{Default is TRUE. Return the raw correlation matrix}

\item{return_plot}{Default is TRUE.  Makes a correlation plot.}

\item{title}{Default is NULL.  Only applies when you are making a plot.}
}
\value{
If both return_raw_numbers and return_plot are true then it will return a list of both.  Otherwise just returns what was indicated.
}
\description{
Correlation function.
}
