% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_drop_constant.R
\name{ezr.drop_constant}
\alias{ezr.drop_constant}
\title{Drop Constants}
\usage{
ezr.drop_constant(dataset, return_n_distinct_tbl = FALSE,
  exclude_cols = NULL, include_NAs = TRUE)
}
\arguments{
\item{dataset}{Dataframe}

\item{return_n_distinct_tbl}{return a 2nd dataframe just showing counts of n_distincts in each column}

\item{exclude_cols}{Exclude this column...use if want to keep a constant}

\item{remove_NAs}{if TRUE, NAs are not counted.  If FALSE, NAs are counted as a unique value}
}
\value{
Returns the corrected dataframe or corrected_dataframe + a n_distinct table
}
\description{
Drop constant columns in a dataframe.  Optional parameter to return 2nd dataframe of counts.
}
\examples{

mtcars$constant = 'constant_value'
ezr.drop_constant(mtcars, return_n_distinct_tbl = FALSE)

}
