% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_filter_null_columns.R
\name{ezr.filter_null_columns}
\alias{ezr.filter_null_columns}
\title{Title Drop Columns in a dataset with excessive NULLs}
\usage{
ezr.filter_null_columns(dataset, threshold_to_eliminate = 0.9,
  keep_columns = NULL, return_only_corrected_df = FALSE)
}
\arguments{
\item{dataset}{Dataframe}

\item{threshold_to_eliminate}{Default is columns 90% and up.}

\item{keep_columns}{Columns to keep even if they have too many null values.  Make sure spelling is accurate.  Should be a vector}

\item{return_only_corrected_df}{Default is FALSE. If this is TRUE then only returns a dataframe}
}
\value{
Returns corrected dataframe
}
\description{
Drop columns in a dataset with columns greater than or equal to a given percentage.
}
