% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_find_columns_containing.R
\name{ezr.find_columns_containing}
\alias{ezr.find_columns_containing}
\title{Find Columns that contain particular values...such as Find Binary Columns}
\usage{
ezr.find_columns_containing(dataset, values = c(0, 1),
  allow_null = TRUE, any_or_only = "only", return_only_true = TRUE)
}
\arguments{
\item{dataset}{dataframe}

\item{values}{the values that you want to check for.  Default is 0 and 1 (checking for binary columns)}

\item{allow_null}{Should NULL be allowed? Default is TRUE.  This means a condition will be met if NA is found in the column.}

\item{any_or_only}{ANY  = does it contain only these values?  Only:  Does it contain only these values?}

\item{return_only_true}{Return a vector of true columns or all columns with trues and falses.}
}
\value{
vector of columns if return_only_true is TRUE or a dataframe if return_only_true is ALSE
}
\description{
Often you want to find columns that are all 0s and 1s...or something else like Y and N.
This function allows you to determine which columns contain the values you are after.  Default is 0 and 1.  There is a flag on whether you want to allow NULL for considering the condition to be met.
}
\examples{
#ezr.find_columns_containing(dataset = mtcars, values = c(0,1), allow_null=FALSE, any_or_only ='only', return_only_true = TRUE)
}
