% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_fix_columnames.R
\name{ezr.fix_colnames}
\alias{ezr.fix_colnames}
\title{Fix column names - replace spaces and periods}
\usage{
ezr.fix_colnames(vector, periods2underscore = TRUE,
  spaces_to_underscore = TRUE, lowercase_all = TRUE,
  uppercase_all = FALSE)
}
\arguments{
\item{vector}{vector of things you want to change}

\item{periods2underscore}{True means replace with underscore.  False means removal.}

\item{spaces_to_underscore}{True means replace with underscore.  False means removal.}

\item{lowercase_all}{Fix the case of the data...Make it all lowercase for ease of use}

\item{uppercase_all}{Fix the case of the data...Make it all uppercase for ease of use}
}
\description{
Replace periods and spaces in column names with underscores
}
