% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_fix_posxct.R
\name{ezr.fix_POSIXCT}
\alias{ezr.fix_POSIXCT}
\title{Deal with POSIXCT fields}
\usage{
ezr.fix_POSIXCT(dataset, retain_time = FALSE, specific_fields = NULL,
  exclude_field = NULL, timezone = "America/Chicago")
}
\arguments{
\item{retain_time}{Do you want to keep the HH:MM:SS in the dataset?  Default is no.}

\item{specific_fields}{Only convert this field}

\item{exclude_field}{Do not convert this field}

\item{timezone}{Default time Zone Is America/Chicago.  Change to NULL to ignore.}
}
\value{
Dataset with POSIXCT fields fixed.
}
\description{
Convert POSIXCT fields to dates
}
