% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_gainslift.R
\name{ezr.gainslift}
\alias{ezr.gainslift}
\title{Gains, Lift, and KS Calculator}
\usage{
ezr.gainslift(df, binary_target, prediction, higher_morelikely = TRUE,
  round_value = 6)
}
\arguments{
\item{df}{Dataframe of prediction results}

\item{binary_target}{Should work for either numeric or factor.  Must be 0s and 1s though.}

\item{prediction}{The prediction value.   Higher values = higher chance of obtaining a 1}

\item{higher_morelikely}{Default is TRUE.  If a higher prediction value is more likely to be a 1.  Set to false if higher is going to be a 0.}

\item{round_value}{Default is 2.  Rounds values to 2 decimal places for a nicer user experience.}
}
\value{
Returns a dataframe with various metrics at each progression into the scorecard.
}
\description{
Outputs a dataframe that has the cumulative capture rate, gains, and KS metrics at each 1% increment of the score.   Also produces the lower threshold of the score.
}
\examples{
ezr.gainslift(df = dataset_telco_churn_from_kaggle, binary_target = 'Churn', prediction ='TotalCharges',higher_morelikely = FALSE )
}
