% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_clf_metrics_thresholds.R
\name{ezr.get_clf_metric_table}
\alias{ezr.get_clf_metric_table}
\title{Get Threshold Classification Metric Table}
\usage{
ezr.get_clf_metric_table(dataset, binary_response, predictor)
}
\arguments{
\item{dataset}{dataframe}

\item{binary_response}{the target/truth/response}

\item{predictor}{the numerical predictor}
}
\value{
A dataframe with all the various classification metrics.
}
\description{
Get the classification metric table from H2O's   metrics$thresholds_and_metric_scores.  This has all of the various TPR, FPR, accuracy, F1, etc metrics.   
The inputs are a binary response of 0 and 1 and a numerical predictor.  It returns a dataframe.
}
\details{
The resulting dataframe and then be plotted as needed.
}
