% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_get_col_index.R
\name{ezr.get_col_index}
\alias{ezr.get_col_index}
\title{Get Column Index positions}
\usage{
ezr.get_col_index(dataset, specific = NULL)
}
\arguments{
\item{dataset}{dataframe}

\item{specific}{Null by default.  Enter a specific column if you need it}
}
\description{
Get all the index positions of each column in a dataframe. You can also get a specific index value
}
\examples{
ezr.get_col_index(dataset=mtcars)


}
