% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_get_date_info.R
\name{ezr.get_date_info}
\alias{ezr.get_date_info}
\title{Get Date Information}
\usage{
ezr.get_date_info(dataset, column)
}
\arguments{
\item{dataset}{A dataframe}

\item{column}{the column that is in date format.}
}
\value{
Returns the original dataset with the date columns added
}
\description{
Quickly get a bunch of different date information about a given column
}
\examples{


}
