% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_get_dummies.R
\name{ezr.get_dummies}
\alias{ezr.get_dummies}
\title{Get dummy encoded columns}
\usage{
ezr.get_dummies(dataset, dummy_this_col, return_as_seperate_df = FALSE,
  drop_original_col = TRUE, convert_to_factor = TRUE)
}
\arguments{
\item{dataset}{Dataframe}

\item{dummy_this_col}{Should be a categorical value you want to one hot encode}

\item{return_as_seperate_df}{Return the original or just the encoded values?}

\item{drop_original_col}{Drop original column? Default is TRUE}

\item{convert_to_factor}{Convert to factor?  Default is TRUE.  Keep at FALSE if you want to later calculate counts or percents of a given level.}
}
\description{
Get one column for each level in a categorical column
}
