% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_get_dummies_all.R
\name{ezr.get_dummies_all}
\alias{ezr.get_dummies_all}
\title{Make all categorical columns one-hot-encoded}
\usage{
ezr.get_dummies_all(dataset, exclude_cols = NULL,
  convert_to_factor = FALSE)
}
\arguments{
\item{dataset}{a dataframe}

\item{exclude_cols}{What columns should this not apply to?}

\item{convert_to_factor}{Should the new columns be converted to factor columns or stay as numeric 0s and 1s}
}
\value{
Returns adjusted dataframe
}
\description{
Convience wrapper for running get dummies on an entire dataset.   Useful when trying to understand the various breakdowns % of a given level.
}
\examples{
ezr.get_dummies_all(iris)
}
