% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_get_object_as_string.R
\name{ezr.get_object_as_string}
\alias{ezr.get_object_as_string}
\title{Object to String}
\usage{
ezr.get_object_as_string(r_object)
}
\arguments{
\item{r_object}{An R object such as dataframe or a variable}
}
\value{
Returns the string of the r object.
}
\description{
Useful for when you want to save a file as its own name or concatenate it with a date...
}
\examples{
myfunct_get_object_as_string(mtcars)
}
