% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_group_infreq.R
\name{ezr.group_infreq}
\alias{ezr.group_infreq}
\title{Group Infrequent Values}
\usage{
ezr.group_infreq(dataset, column, le_percent = 0.05,
  newname = "__infreq__", change_single = FALSE, convert_na = TRUE,
  factorize_new_column = TRUE)
}
\arguments{
\item{dataset}{Dataset}

\item{column}{The column you are interested in re-grouping the infrequent values of.}

\item{le_percent}{The <= percent used to change values to <newname>}

\item{newname}{What it should be called.  Defaults to '__infreq__'}

\item{change_single}{If there are just two values in a column, should anything be changed? Default is no.}

\item{convert_na}{Should NA values be converted to factors?  Default is TRUE.}

\item{factorize_new_column}{Convert the new column to factors?  default is true}
}
\value{
Returns the dataframe that was inputted with adjustments made, if needed
}
\description{
Ideally only use this for categorical values - however there is no check or prevention against using it with numericals
}
\examples{
newdata=ezr.group_infreq(diamonds, column = 'color', le_percent = 0.20)
}
