% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_glimpse.R
\name{ezr.h2o.glimpse}
\alias{ezr.h2o.glimpse}
\title{H2O Version of Glimpse}
\usage{
ezr.h2o.glimpse(h2odataset, n_head_values = 0)
}
\arguments{
\item{h2odataset}{h2odataframe}

\item{n_head_values}{Number of values to concatenate into a comma seperated list. Keep at 0 to turn off}
}
\value{
dataframe of column name, type, and col-index position.  Onptionally returns the first N rows
}
\description{
H2o's version of glimpse.  You control the <N> to be shown from header
}
