% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr.h2o_basic_filter.R
\name{ezr.h2o_basic_filter}
\alias{ezr.h2o_basic_filter}
\title{H2o Subset Rows}
\usage{
ezr.h2o_basic_filter(h2o_df, column_to_filter, sign, value)
}
\arguments{
\item{h2o_df}{h2odataframe}

\item{column_to_filter}{name of column to filter using}

\item{sign}{should be >=, >, <, <=, ==, !=.   In future will add null/not null}

\item{value}{what value is being compared.}
}
\value{
H2o dataframe after filtering is applied
}
\description{
Convience wrapper extremely simple subsets in h2o.
 Provides the number of records returned as a result of filter.
}
