% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_factor.R
\name{ezr.h2o_factor_fix}
\alias{ezr.h2o_factor_fix}
\title{Factorize Variabels in h2o}
\usage{
ezr.h2o_factor_fix(h2oframe, variables)
}
\arguments{
\item{h2oframe}{the h2o dataframe}

\item{variables}{the variables you wish to factor}
}
\value{
Returns the modified h2o dataframe
}
\description{
Sometimes you when you load an h2o frame the datatypes aren't respected when coming from R - especially 0s and 1s.
Or you have a list of variables you want to factor...
}
\details{
This handles that process for you.
}
