% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_grid_gbm.R
\name{ezr.h2o_gbm_grid}
\alias{ezr.h2o_gbm_grid}
\title{GBM Grid Search}
\usage{
ezr.h2o_gbm_grid(train_df, valid_df = NULL, xvars = names(train_df),
  yvar = "target", grid_id = "gbm_grid", prescreengbm = TRUE,
  prescreen_keepvars_criteria = "percent",
  prescreen_keepvars_threshold = 0.005, xval = TRUE, folds = 5,
  keep_cross_validation_predictions = TRUE, max_models = 1,
  learnrate = 0.025, max_min_runtime = 15, ntrees = 125,
  seed = 2018, ...)
}
\arguments{
\item{train_df}{h2o dataframe}

\item{valid_df}{a validation dataframe.  Default is NULL.  If NULL it the train_df will be split into 80/20 split and the 20% will be used for validating to guard against overfit}

\item{xvars}{default is everything in training df}

\item{yvar}{target}

\item{grid_id}{grid id to use.  Default is gbm_grid}

\item{prescreengbm}{Default is TRUE.  Should a pre-screen be run to eliminate excess variables?  This will run a gbm with default params, and be used to eliminate variables before re-training.  This is to prevent against 100s of variables with 0.001 or similar importance criteria in model.}

\item{prescreen_keepvars_criteria}{Valid values are 'percent' and 'number' Default is 'percent' importance.  Number refers to how many variables such as 5/10/100}

\item{prescreen_keepvars_threshold}{Default threshold is 0.01 for percent for retention.  Enter an integer for 'count'.  If the value is <= 1 and the <prescreen_keepvars_criteria> is equal to 'number' then this will default to 25.}

\item{xval}{Default is TRUE.}

\item{folds}{Default is 5}

\item{keep_cross_validation_predictions}{Default is FALSE}

\item{max_models}{Default is 1.  If value is 1, then a default GBM will run}

\item{learnrate}{Default is 0.05. You can enter a vector c(0.01, 0.05)}

\item{max_min_runtime}{How many minutes can this run for?  Default is 15min}

\item{ntrees}{Default is 100.}

\item{seed}{Default is 2018}

\item{...}{Additional inputs...}
}
\value{
Returns a grid of models
}
\description{
Off the shelf grid search for GBM w/ hyper parameters.  Initial parameters
}
\examples{
library(h2o)
h2o.init()
h2odf = as.h2o(dataset_telco_churn_from_kaggle)
example_grid_search=ezr.h2o_gbm_grid(train_df = h2odf, yvar='Churn', max_models = 11)
}
