% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_get_grid.R, R/h2o_get_grid.R
\name{ezr.h2o_get_grid}
\alias{ezr.h2o_get_grid}
\title{Get H2o Grid}
\usage{
ezr.h2o_get_grid(grid, decreasing = TRUE, sort_by = "auc",
  classification = TRUE, rounding_digits = 3,
  return_best_model = TRUE, return_vector_model_ids = TRUE,
  print = TRUE)

ezr.h2o_get_grid(grid, decreasing = TRUE, sort_by = "auc",
  classification = TRUE, rounding_digits = 3,
  return_best_model = TRUE, return_vector_model_ids = TRUE,
  print = TRUE)
}
\arguments{
\item{grid}{Grid id or grid object}

\item{sort_by}{sort by metric. Valid values are 'auc','f1','mcc','logloss','recall','precision','lift_top_group','accuracy','gini_coef', 'mse','rmse','r2'}

\item{classification}{Default is True.  Need to tell function if you want to return appropriate set of metrics}

\item{rounding_digits}{How many digits to round results}

\item{sort_higher_first}{Default is true.  Higher=Better}

\item{grid}{Grid id or grid object}

\item{sort_higher_first}{Default is true.  Higher=Better}

\item{sort_by}{sort by metric. Valid values are 'auc','f1','mcc','logloss','recall','precision','lift_top_group','accuracy','gini_coef', 'mse','rmse','r2'}

\item{classification}{Default is True.  Need to tell function if you want to return appropriate set of metrics}

\item{rounding_digits}{How many digits to round results}
}
\value{
Dataframe of the models from the grid search and the metrics.

Dataframe of the models from the grid search and the metrics.
}
\description{
Improves upon standard h2o.getGrid.  Returns more metrics at once.   Remember to change DECREASING if the value is better higher/lower.

Improves upon standard h2o.getGrid.  Returns more metrics at once.   Remember to change DECREASING if the value is better higher/lower.
}
