% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_get_gridmodels.R
\name{ezr.h2o_get_gridmodels}
\alias{ezr.h2o_get_gridmodels}
\title{Get H2o Grid/AutoMl Model IDs}
\usage{
ezr.h2o_get_gridmodels(h2o_grid)
}
\arguments{
\item{h2o_grid}{Doesn't matter if string or model object.  Can be either an h2o grid or h2o-automl}
}
\value{
Returns a vector of model ids so you can use these in a loop
}
\description{
Get H2o Grid/AutoMl Model IDs
}
