% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_get_thresholds.R
\name{ezr.h2o_get_thresholds}
\alias{ezr.h2o_get_thresholds}
\title{Get Threshold Metrics from h2o classification models}
\usage{
ezr.h2o_get_thresholds(h2omodels, xval = TRUE, newdata = NULL,
  valid = FALSE, train = FALSE)
}
\arguments{
\item{h2omodels}{a vector of models such as c('id1','id2','id3')}

\item{xval}{Use cross validation metrics?.  This is true by default.}

\item{newdata}{Use test data? Provide the test dataframe here}

\item{valid}{Use validation data?}

\item{train}{Use the training data?}
}
\value{
Returns a dataframe of the concatenated metrics along with the model-id and algo.
}
\description{
Returns the entire table of threshold metrics for each model.  Use the ezr.h2o_get_gridmodels to get the modelids out of grids.    Determine whether or not these metrics should be at the training, xval, valid, or test data(newdata) level.
}
\details{
Only one metric should be true at a time
}
\examples{
  ezr.h2o_get_thresholds(h2omodels = your_vector_of_models, xval=FALSE, newdata='your_test_h2oframe' )
}
