% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_top_vars.R
\name{ezr.h2o_get_top_vars}
\alias{ezr.h2o_get_top_vars}
\title{Get Top Variables from Model}
\usage{
ezr.h2o_get_top_vars(model, n = 20, return_as_vector = TRUE)
}
\arguments{
\item{model}{h2o model object or string model id.}

\item{n}{top n variables.  20 by default}

\item{return_as_vector}{True by default}
}
\value{
Vector or dataframe of top N variables.
}
\description{
Get the top N variables from a model.  Convienence wrapper.
}
