% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_grid_glm.R
\name{ezr.h2o_glm_grid}
\alias{ezr.h2o_glm_grid}
\title{GLM Grid}
\usage{
ezr.h2o_glm_grid(train_df, valid_df = NULL, xvars = names(train_df),
  no_regularization = FALSE, family = "binomial", yvar = "target",
  grid_id = "glm_grid", use_prescreen = TRUE,
  prescreen_keepvars_criteria = "number",
  prescreen_keepvars_threshold = 30, xval = TRUE, folds = 5,
  keep_cross_validation_predictions = FALSE, max_models = 4,
  max_min_runtime = 15, ...)
}
\arguments{
\item{train_df}{Training dataframe}

\item{valid_df}{If not provided, the training dataframe is split for you 80/20}

\item{xvars}{The xvariables in the model}

\item{no_regularization}{Default is FALSE.  Set to TRUE, if you want to run plain logistic / linear regression.}

\item{family}{.  Default is 'binomial'.   Use binomial for classification with logistic regression, others are for regression problems. Must be one of: "gaussian", "binomial", "quasibinomial", "ordinal", "multinomial", "poisson", "gamma", "tweedie".}

\item{yvar}{The target variable}

\item{grid_id}{Name of Grid ID}

\item{use_prescreen}{Default to TRUE}

\item{prescreen_keepvars_criteria}{Default is number, 30.  Valid values here are 'percent' or 'number'.  Number picks a certain number of variables, percentage picks a percentage from xgboost model.}

\item{prescreen_keepvars_threshold}{Default is 30.  Change this to percentage if you want percentages}

\item{xval}{Cross validation, TRUE/FALSE}

\item{folds}{# of Folds if you use cross validation}

\item{keep_cross_validation_predictions}{, Keep the predictions? Defaults to false.}

\item{max_models}{Defaults to 1}

\item{max_min_runtime}{Defaults to 15min.  Remember to enter this as minutes, not seconds.}

\item{...}{Additional model parameters.}

\item{seed}{Defaults to 2018}
}
\value{
A grid searched models
}
\description{
Run GLM grid search against alphas and lambdas.  Additionally, there is the option to pre-screen with an xgboost model to narrow down importance features.
}
\details{
Models can be run with Xval, and use different GLM types.

Additioinally, there is an easier interface for running regular linear/logistic regressions without any regularization.

Extra parameters can be added with ... in particular, the max_active_predictors, should be added to help obtain sparse solutions, especially when you don't want model to run forever.

Note that it is ideal to setup the train and valid dataframes before passing in the data so you are certain of what they are.  Note, that for non-regularization models,
}
