% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2omerge_preds.R
\name{ezr.h2o_merge_preds}
\alias{ezr.h2o_merge_preds}
\title{Merge H2o Predictions}
\usage{
ezr.h2o_merge_preds(model, dataset, xval = TRUE, as_dataframe = FALSE,
  nameof_pred_column = "prediction", other_fields_to_retain = NULL)
}
\arguments{
\item{model}{model.  Should be binary or regresison model}

\item{dataset}{dataframe either test dataset or xval dataset}

\item{xval}{XVAL=TRUE is default.}

\item{as_dataframe}{Convert h2o dataframe to dataframe}

\item{nameof_pred_column}{Default is 'prediction'}

\item{return_2cols}{Return just the target and the prediction value?  Default is TRUE}
}
\description{
Merge predictions from xval or regular predictions on a test dataset.
}
