% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_outliers.R
\name{ezr.h2o_outliers}
\alias{ezr.h2o_outliers}
\title{Anomaly Detection - Deep Learning & Isolation}
\usage{
ezr.h2o_outliers(dataset, y = NULL, x = NULL, deep_learning = FALSE,
  isolation_forest = TRUE, hidden = c(10, 10), epochs = 100,
  return_as_data_frame = TRUE, ntrees = 75, max_depth = 9,
  max_runtime_min = 5, return_extras = TRUE)
}
\arguments{
\item{dataset}{Dataset}

\item{y}{The target.  You don't want to use this with the X variables used to identify anamolies.}

\item{x}{Variables to consider for anamoly detection.}

\item{deep_learning}{Default is FALSE.  Both this and Isolation forest can be TRUE, to see side-by-side comparison/agreement.}

\item{isolation_forest}{Default is TRUE.  Use this algo to detect anamolies.}

\item{hidden}{DL parameter}

\item{epochs}{DL parameter}

\item{ntrees}{Default is 75.}

\item{max_depth}{Default is 9.  May wish to increase if having trouble identifying anamolies - if the max seperation is close to this number.}

\item{max_runtime_min}{Time in minutes for each type of model to run}
}
\value{
Original Dataframe, with outlier columns added: <outlier_dl> and <outlier_iso_forest>.  Higher values are larger outliers
}
\description{
Find outlier datapoints in a dataset and return entire dataframe with the following columns if the models were run:
}
\details{
For isolation forest: anomaly_iso_predict, anomaly_iso_length, pct_rank_iso_anomaly
For DL model:   anomaly_dl_recon_mse, pct_rank_dl_anomaly

Lower Ranks such as 0.01 indicate something that very likely is an outlier.   Higher values for predict indicate that somethign is an outlier.   anomaly_iso_length refers to mean number of splits from the isolation forest to classify as an outlier.
}
