% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_viewtree.R
\name{ezr.h2o_viewtree}
\alias{ezr.h2o_viewtree}
\title{Print H2o Tree}
\usage{
ezr.h2o_viewtree(mojo_location, print_tree = 0, file_name = "mymodel")
}
\arguments{
\item{mojo_location}{Should be the path + file name ending in .zip.  For example "Users/yourname/Desktop/xgboost_model.zip"}

\item{print_tree}{Use 0 for the first one}

\item{file_name}{Name of the png file to create.  Note it may be hard to read it for big trees.  Opens in Chrome.}

\item{h2o_model}{NULL by default.  If you haven't saved a mojo file yet you can get to it from R environment.  This}

\item{location_of_h2o_jar}{null by}
}
\value{
Returns a plot of the tree. It may be hard to read since trees can be hard to view.  Best place to look at this is in chrome.
}
\description{
View the splitting of a decision tree in an image.
}
