% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_xgb_importance_aggregated.R
\name{ezr.h2o_xgb_importance_aggregated}
\alias{ezr.h2o_xgb_importance_aggregated}
\title{Aggregate Importance by Level in H2o}
\usage{
ezr.h2o_xgb_importance_aggregated(model)
}
\arguments{
\item{model}{Should be a string of model name}
}
\value{
Returns importance metrics, with categoricals aggregated...
}
\description{
Aggregate importance for xgboost categorical values into a single variable, just like with gbm or random forest.  If you want to see the true importance of a given level of the categorical value, use the regular importance method.   The 'type' column in the result will show if it is numeric or categorical, and if categorical, how many levels are present..
}
