% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2ovar_compare.R
\name{ezr.h2ovar_compare}
\alias{ezr.h2ovar_compare}
\title{Compare Variable Importance Across Models}
\usage{
ezr.h2ovar_compare(h2o_models)
}
\arguments{
\item{h2o_models}{Should be a vector of model names.....  h2o_models= as.data.frame(your_grid_of_h2o_models<atsign>summary_table)$model_ids}
}
\value{
Returns a list of two dataframes.  One by algo and and one by all models.
}
\description{
Fixes needed for when GLMs are included... currently only works with DRF, GBM, XGB
}
