% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_impute.R
\name{ezr.impute}
\alias{ezr.impute}
\title{Impute values}
\usage{
ezr.impute(dataset, use_mean = FALSE, use_median = TRUE,
  only_columns = NULL, adjust_chars = FALSE, exclude_columns = NULL)
}
\arguments{
\item{dataset}{Dataset}

\item{use_mean}{Impute with the mean value?}

\item{use_median}{Impute with the median value?  This is the default value}

\item{only_columns}{Just look at these specific columns}

\item{adjust_chars}{Should character/factor columns be imputed with mode?  Default is FALSE}

\item{exclude_columns}{Do not adjust these columns at at all.}
}
\description{
Impute values in your dataframe on missings.  Pick either mean or median for numericals. Mode is used for categoricals.
}
