% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_json_fix.R
\name{ezr.json_to_df}
\alias{ezr.json_to_df}
\title{Json to DF}
\usage{
ezr.json_to_df(dataset, json_column)
}
\arguments{
\item{dataset}{dataframe}

\item{json_column}{Column that contains the json}
}
\description{
Flip a json column to a dataframe
}
