% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_make_ids.R
\name{ezr.make_ids}
\alias{ezr.make_ids}
\title{Title Create RowID by variables or for each row}
\usage{
ezr.make_ids(dataset, unique_by_row = TRUE, grouping_fields = NULL,
  id_field_name = "myid")
}
\arguments{
\item{dataset}{dataset}

\item{unique_by_row}{TRUE by default.  Each record will be numbered 1...nrow(dataset)}

\item{grouping_fields}{Set unique_by_row to FALSE.  This creates a group-by-variable ID}

\item{id_field_name}{What should the new column be called.  Default is 'myid'}
}
\value{
Returns a dataframe with ID added.
}
\description{
Get an ID for each row or for each combination of values across columns.
}
