% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_map_datatypes.R
\name{ezr.map_datatypes}
\alias{ezr.map_datatypes}
\title{Convert datatypes by lookup}
\usage{
ezr.map_datatypes(dataset, lookup_dataset)
}
\arguments{
\item{lookup_dataset}{Should be a dataframe of two columns.  Each row is a variable.   datatype values are:     character or char /   as.factor or factor / numeric or number or float / date or as.date}

\item{target_dataset}{The dataframe that contains your data}
}
\value{
Returns the dataframe with corrected datatypes.
}
\description{
Convert datatypes in a dataframe by a lookup table.
}
