% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_mode.R
\name{ezr.mode}
\alias{ezr.mode}
\title{Mode}
\usage{
ezr.mode(dataframe, column, single_mode = FALSE)
}
\arguments{
\item{dataframe}{dataframe}

\item{column}{What column to calculate mode for}

\item{single_mode}{Default is FALSE, which returns all values even if there is a tie.}
}
\description{
Get the mode for a column in a dataframe
}
