% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_plot_multi_lines.R
\name{ezr.multi_line_plot}
\alias{ezr.multi_line_plot}
\title{Multiple Line Plots Same X Axis}
\usage{
ezr.multi_line_plot(dataset, x, y1, y2 = NULL, grouping = NULL,
  title = NULL, y_lab = NULL, date_breaks = NULL,
  date_labels = NULL)
}
\arguments{
\item{dataset}{dataframe}

\item{x}{normally a date, but can be continous}

\item{y1}{y axis value}

\item{y2}{NULL by default.  Use if the value to be plotted is in a 2nd column}

\item{grouping}{NULL by default.  Use if the value to be plotted is dependant upon a group}
}
\value{
A plot
}
\description{
Plots multiple lines against the same x axis.  Often the X value will be a date field.
}
\details{
Option A:  If the numerical value to plot is in two columns use y1 and y2.
 Option B:  If the numerical value to plot is in single column, but to be plotted by a grouped column, use the 'grouping' field
}
