% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_auc_plot.R
\name{ezr.plot_auc}
\alias{ezr.plot_auc}
\title{AUC PLOT}
\usage{
ezr.plot_auc(dataset, fpr_or_truth, tpr_or_pred, pred_2nd_model = NULL,
  title = NULL, metrics_precalculated = FALSE, add_text = TRUE)
}
\arguments{
\item{dataset}{dataset}

\item{fpr_or_truth}{fpr value or truth.  Default is expecting a 0/1 truth value.}

\item{tpr_or_pred}{fpr or prediction value.  Default expects a prediction value.}

\item{pred_2nd_model}{Default is NULL.  If you wish to compare two plots side-by-side, then you must use truth & prediction values.}

\item{title}{plot title}

\item{metrics_precalculated}{FALSE by default.  This means expects a 0/1 truth_or_fpr and a score in the pred_or_tpr.  Set to TRUE if you already know that the FPR and TPR are already calculated for you.}

\item{add_text}{add text to plot?  Default is true.}
}
\value{
Returns a plot of the AUC-values
}
\description{
Generates an AUC plot.  Can compare two different fields against each other if needed.   Can plot either FPR vs TPR columns or   Truth vs prediction value(s).  Adds the AUC alues to the plots.
}
