% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_plot_clf_metrics.R
\name{ezr.plot_clf_metrics}
\alias{ezr.plot_clf_metrics}
\title{Plot Various Classification Metrics at Once}
\usage{
ezr.plot_clf_metrics(dataset, truth, prediction,
  higher_more_likely = TRUE)
}
\arguments{
\item{truth}{The truth variable}

\item{higher_more_likely}{Higher prediction values correspond with greater chance of 1 for the truth}

\item{target}{The target variable}
}
\description{
Plot various classification metrics:
}
\details{
1.  Basic distribution and counts
2.  AUC
3.  PRAUC
4.  Cum Lift
5.  Cum Capture Rate
6.  Cum Gain (n) over a random guess
}
\examples{
ezr.plot_clf_metrics(dataset = easyr::dataset_telco_churn_from_kaggle, truth = 'Churn','TotalCharges',higher_more_likely = FALSE)
}
