% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_plot_facet_numericals.R
\name{ezr.plot_facet_numericals}
\alias{ezr.plot_facet_numericals}
\title{Facet Plot Numericals}
\usage{
ezr.plot_facet_numericals(dataset, variables_to_plot = names(dataset),
  type = "box", grouping_value, sample_frac = 1)
}
\arguments{
\item{dataset}{Dataframe}

\item{variables_to_plot}{vector of numerical columns.  Only plotting of numericals is allowed}

\item{type}{'box' or 'density'.  Box is default}

\item{grouping_value}{The grouping value.  Normally this is the binary 0/1 target}
}
\value{
a plot
}
\description{
Facet plot numerical values in a dataframe by a group - typically your target, but can be any categorical value.  Avoid extremely large number of plots
}
\examples{
ezr.plot_facet_numericals(dataset = mtcars, grouping_value = 'vs')
}
