% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_plot_gainslift_compare.R
\name{ezr.plot_gainslift_compare}
\alias{ezr.plot_gainslift_compare}
\title{Compare Gainslift Using Multiple Predictors}
\usage{
ezr.plot_gainslift_compare(dataset, binary_target, fields,
  higher_morelikely = TRUE, plot_type = "capture", return_tbl = TRUE,
  y_title = NULL, plot_title = NULL)
}
\arguments{
\item{dataset}{The dataset.  Can be an h2o dataframe}

\item{binary_target}{the target.  Should be 0/1.}

\item{fields}{The fields that}

\item{higher_morelikely}{Is a higher value more likely of a value of '1'?  Default is TRUE.   If multiple columns are provided, it is assumed they all have the same direction.   If they don't have the same direction, manipulate the data prior to using this function.}

\item{plot_type}{Default is 'capture'.   Valid values are 'capture', 'response', 'gains', or 'lift'.   All of these values are cumulative.}

\item{return_tbl}{Return the dataframe of values?}

\item{plot_title}{= Name of plot}
}
\description{
Compare and plot  Cumulative Capture Rate, Gains, Lift, or Response Rate across 1 or more predictors.   Optionally returns the underlying dataframe.
}
