% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_plot_vs_time.R
\name{ezr.plot_vs_time}
\alias{ezr.plot_vs_time}
\title{Plot a value over time}
\usage{
ezr.plot_vs_time(dataset, date_field, value, date_grouping = "month",
  plt_metric = "mean", plot_type = "line", show_text = TRUE,
  title = NULL)
}
\arguments{
\item{dataset}{dataframe}

\item{date_field}{The date field to plot against}

\item{value}{If the field is a character or}

\item{date_grouping}{How to group data for plotting. Data is always 'floored'...that is 2018-12-25 would be rounded to 2018-12-01.  Default is MONTH.  Options are month, week, day, hour , year, quarter,season,halfyear and none.}

\item{plt_metric}{Plotting metric. Valid values are count, median, mean, sd, max,min, count1.  'mean' is default.   Count1 is valid only for categoricals and assumes a 0/1 value.}

\item{title}{Plot title.  Default is nothing.}

\item{plot_type.}{Default is line.  Valid values are 'line','bar', and 'box'.  Box should be used when limited number of grouping values for the date field}

\item{show_text.}{Default is FALSE.  Not yet implemented...}
}
\value{
Returns a plot
}
\description{
Plot a value over time.   You control the date grouping.  See params for valid  values.    For example how many records by date, or the mean vaue by date.   If using a 0/1 column, you can use mean to get the % 1s.
}
\details{
Bar, Box, or Line Plots are possible.   Default is line.
}
