% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_quantiles.R
\name{ezr.quantiles}
\alias{ezr.quantiles}
\title{Quantiles}
\usage{
ezr.quantiles(dataset, exclude_columns = NULL, only_columns = NULL,
  number_quantiles = 20)
}
\arguments{
\item{dataset}{dataframe}

\item{exclude_columns}{dont use these columns.  should be a vector or in 'quotes'}

\item{only_columns}{use only these columns.  should be a vector or in 'quotes'}

\item{number_quantiles}{20 will get you every 5%tile, 100 will get you 0-100, etc.  Default is 20}
}
\description{
Quantiles
}
\examples{
ezr.quantiles(mtcars, number_of_quantiles = 10)
}
