% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_range01.R
\name{ezr.range01}
\alias{ezr.range01}
\title{Transform a value from range 0 to 1}
\usage{
ezr.range01(var)
}
\arguments{
\item{var}{the variable that you want to transform.}
}
\value{
Returns the transformed values
}
\description{
Easily rescale a value from 0 to 1.
}
\examples{

ezr.range01(mtcars$hp)
ezr.range01(mtcars \%>\% select(hp))

}
