% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_refresh_info_tbl.R
\name{ezr.refresh_info_tbl}
\alias{ezr.refresh_info_tbl}
\title{Get ezr function description in table}
\usage{
ezr.refresh_info_tbl(workbook_name = "ezr_function_lookup")
}
\arguments{
\item{workbook_name}{Defaults to my ezr lookup table}
}
\value{
Returns a dataframe of the google worksheet.
}
\description{
Get a list of functions available in the table for easy filtering.
Can also be used to load a google worksheet file.
You may need to run gs_auth() first.
}
\examples{
 function_lookups = ezr.refresh_info_tbl()
}
