% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr.regr_eval.R
\name{ezr.regr_eval}
\alias{ezr.regr_eval}
\title{Regression Evaluation Summary}
\usage{
ezr.regr_eval(dataset, y_true, y_pred)
}
\arguments{
\item{dataset}{Dataframe}

\item{y_true}{Truth value}

\item{y_pred}{Predicted Value}
}
\value{
Returns a dataframe with many metrics on the comparison of the two numbers
}
\description{
Get a variety of metrics around two numerical vectors.
}
