% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_rename_by_lookup.R
\name{ezr.rename_by_lookup}
\alias{ezr.rename_by_lookup}
\title{Title Rename Columns by Lookup}
\usage{
ezr.rename_by_lookup(df, df_column_lookup)
}
\arguments{
\item{df}{The datafraine}

\item{df_column_lookup}{a dataframe with a new and old column for the spelling of names.  It should have a 'new' and an 'old' column name.}
}
\description{
Rename columns in a dataframe by a lookup table.
}
\details{
Lookup table must have the column names 'old' and 'new'
}
