% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_replace.R
\name{ezr.replace}
\alias{ezr.replace}
\title{Replace Values in a Dataframe}
\usage{
ezr.replace(dataset, existing_value, replacement_value,
  comparison_sign = "equal", use_fields = NULL,
  exclude_fields = NULL)
}
\arguments{
\item{dataset}{Dataframe}

\item{existing_value}{The existing value.}

\item{replacement_value}{The replacement value.}

\item{comparison_sign}{Equal, greater_than, less_than, are valid values}

\item{fields}{Default is to check and replace in every column.}
}
\description{
Replace values in a dataframe.  H2odataframes not yet supported.   This should be used for general and universal replacements.  It should not be used for imputing values.  If you wish to use a mean/median/mode replacement function, check out  ezr.impute
}
