% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_split_data.R
\name{ezr.split_data}
\alias{ezr.split_data}
\title{Split Data}
\usage{
ezr.split_data(dataset, perc = c(0.75), strata = NULL,
  return_as_single_df = FALSE, seed = 2019,
  datasplit_identifiers = c("train", "test", "valid"))
}
\arguments{
\item{dataset}{dataset.  H2o or regular.}

\item{strata}{Default is NULL.  This is for startified sampling.  Not valid for h2o dataframes.}

\item{return_as_single_df}{Return as a single dataframe.}

\item{prop}{A vector of values.  Only a single valid is needed for standard train/test split.  If a 2nd value is entered then a valid dataset will be created.}

\item{datasplit_identifiers.}{Assumed to be in this order:  train/test/valid.  You may call things otherwise, but the returned dataset may be named differently.}
}
\description{
Seperate data into train and test datasets.  Data can be returned in a single dataset or in multiple
}
