% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_sql_functions.R
\name{ezr.sql_safe_query}
\alias{ezr.sql_safe_query}
\title{Title Safe Query}
\usage{
ezr.sql_safe_query(query, con = con, limit_rows = 20)
}
\arguments{
\item{query}{The query that you want to run}

\item{con}{The name of the connection to the DB.  Default is 'con'}

\item{limit_rows}{How many rows should be returned?  Default is 20 which prevents a massive query return.}
}
\value{
Returns a dataframe of results
}
\description{
Safely run a query or control the number of records returned with this function.  Use this function when examining or testing queries.
}
