% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_sql_functions.R
\name{ezr.sql_sample_query}
\alias{ezr.sql_sample_query}
\title{Title Sample Table Results}
\usage{
ezr.sql_sample_query(tbl_name, con = con, limit_rows = 20)
}
\arguments{
\item{tbl_name}{The database table name.  An example would be 'customers' or 'sales'}

\item{con}{The connection name.  The default connection name is 'con' that you used when you initiated the connection with the DB.  Change it if you called it something else.}

\item{limit_rows}{Default is 20 rows.  This prevents accidental large queries.}
}
\value{
Returns a dataframe of results
}
\description{
Intent of this function is to examine contents of a database table.  A safety limit of 20 rows is applied by default so you don't accidentally query a massive table.
}
