% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_image_to_table.R
\name{ezr.tbl_to_image}
\alias{ezr.tbl_to_image}
\title{Table to Image}
\usage{
ezr.tbl_to_image(dataset, n_records = 10, only_columns = NULL,
  exclude_columns = NULL, theme = NULL)
}
\arguments{
\item{dataset}{Dataframe to make image of.}

\item{n_records}{Only use n records such as when printing a few rows from dataframe.  Default is 10.}

\item{only_columns}{Default is FALSE. Only use some columns.}

\item{exclude_columns}{Default is FALSE. Exclude some columns.}

\item{theme}{Theme.  See ggpubr::ggtexttable for more.  Default is 'mBlack'.  'classic' is other good option.}
}
\value{
Returns a tableGrob which can be plotted with grid::grid.table().  Auto generates image.
}
\description{
Converts a table to an image.
}
