% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_unquoted_c.R
\name{ezr.unquoted_c}
\alias{ezr.unquoted_c}
\title{Get a quoted vector from an unquoted vector}
\usage{
ezr.unquoted_c(...)
}
\arguments{
\item{...}{a vector of values that do not have quotes.}
}
\value{
A vector of quoted values
}
\description{
The standard function "c" requires quoted values when you are copying/pasting values in.  For example your_vector = c("job", "income","state").  This function
allows for this to be written as your_vector = c(job, income, state) - the same thing without the quotes.  This makes it easier when entering a list of values as you don't have to manually add all the quotes.
}
