% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2fac.R
\name{fac2char}
\alias{fac2char}
\title{Factors to Characters}
\usage{
fac2char(x)
}
\arguments{
\item{x}{Data frame to modify.}
}
\value{
Data frame with converted characters.
}
\description{
Convert all factor columns in a data frame to characters.
Author: Bryce Chamberlain.
}
\examples{
char2fac( iris )
}
