% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_col.R
\name{find_max_col}
\alias{find_max_col}
\title{Takes a data frame and returns a DF with the same columns but with the max(length)
of each column as the only observation}
\usage{
find_max_col(full_df)
}
\arguments{
\item{full_df}{name of data frame to assess}
}
\value{
Data frame described above
}
\description{
Takes a data frame and returns a DF with the same columns but with the max(length)
of each column as the only observation
}
