% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Link data.tables by fuzzy string matching}
\usage{
fuzzy_match(a, b, acol, bcol, method = "jw", ...)
}
\arguments{
\item{a}{a source data.table}

\item{b}{a target data.table}

\item{acol}{column name in \code{a} to use for matching}

\item{bcol}{column name in \code{b} to use for matching}

\item{method}{method for \code{\link[stringdist]{stringdistmatrix}}}
}
\value{
a data.table containing any blocking columns,
the source column,
the closest match in the target column,
and the string distance for that match.
}
\description{
Finds the closest string match between two data.tables.
The default method computes Jaro-Winkler string distances using the \code{stringdist} package.
In cases with multiple closest matches, only the first match is reported.
}
\examples{
library(data.table)

set.seed(575)
DTA <- data.table(block1 = sample(LETTERS[1:4], 20, TRUE),
                  block2 = sample(LETTERS[1:4], 20, TRUE),
                  fruit   = sample(stringr::fruit[1:12], 20, TRUE))

DTB <- data.table(block1 = sample(LETTERS[1:4], 20, TRUE),
                  block2 = sample(LETTERS[1:4], 20, TRUE),
                  fruit   = sample(stringr::fruit[1:12], 20, TRUE))

fuzzy_match(DTA, DTB, "fruit", "fruit")

setkey(DTA, block1, block2)
setkey(DTB, block1, block2)

DTA[ , fuzzy_match(.SD, b = DTB[.BY], "fruit", "fruit"),
       by = .(block1, block2)]

}
\seealso{
\link[stringdist]{stringdistmatrix}
}
