% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_up.R
\name{get_google_drive_path}
\alias{get_google_drive_path}
\title{Get Filestream path}
\usage{
get_google_drive_path()
}
\value{
The OS-specific mounting of Google Drive
}
\description{
Identifies the operative system in which we are running, and returns the correct path to where Google Drive is mounted
}
