% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbetterint.R
\name{getbetterint}
\alias{getbetterint}
\title{Get better Int}
\usage{
getbetterint(int)
}
\arguments{
\item{int}{Vector of character bucket names to transform.}
}
\value{
Vector of transformed values.
}
\description{
Takes bucket names of binned values such as [1e3,2e3) or [0.1234567, 0.2) and formats the values nicely into values such as 1,000-2,000 or 0.12-0.20
Author: Scott Sobel. Tech Review: Bryce Chamberlain.
}
\examples{
iris$bin <- binbyvol( iris, 'Sepal.Width', 'Sepal.Length', 5 )
getbetterint( iris$bin )
}
