% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getinfo}
\alias{getinfo}
\title{Get Info}
\usage{
getinfo(
  df,
  colname,
  topn = 5,
  botn = 5,
  graph = TRUE,
  ordered = TRUE,
  display = TRUE,
  cutoff = 20,
  main = NULL,
  cex = 0.9,
  xcex = 0.9,
  bins = 50,
  col = "light blue"
)
}
\arguments{
\item{df}{Data Frame or Data Table.}

\item{colname}{(Character) Name of the column to get information about.}

\item{topn}{(Optional) Number of top values to print.}

\item{botn}{(Optional) Number of bottom values to print.}

\item{graph}{(Boolean Optional) Output a chart of the column.}

\item{ordered}{(Optional)}

\item{display}{(Optional)}

\item{cutoff}{(Optional)}

\item{main}{(Optional)}

\item{cex}{(Optional)}

\item{xcex}{(Optional)}

\item{bins}{(Optional)}

\item{col}{(Optional)}
}
\value{
Only if display = FALSE, returns information about the column. Otherwise information comes through the graphing pane and the console (via cat/print).
}
\description{
Get information about a Column in a Data Frame or Data Table. Use getdatadict to explore all columns in a dataset instead.
Author: Scott Sobel. Tech Review: Bryce Chamberlain.
}
\examples{
getinfo(iris,'Sepal.Width')
getinfo(iris,'Species')
}
