% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not-In}
\usage{
needle \%ni\% haystack
}
\arguments{
\item{needle}{Vector to search for.}

\item{haystack}{Vector to search in.}
}
\value{
Boolean vector/value of comparisons.
}
\description{
Opposite of %in% operator. From https://stackoverflow.com/a/46867726/4089266.
Author: Bryce Chamberlain.
}
\examples{
c(1,3,11) \%ni\% 1:10
}
