% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not In}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{the value you want to check for}

\item{table}{The list of values you want to see if it is in.}
}
\value{
TRUE or FALSE response
}
\description{
A function that is the opposite of %in%.   You can write this as:
3 %not_in% c(3,4) or    %not_in%(3,c(3,4)).   If you use the 2nd method, you must use the `` before and after the percent signs.
}
