% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashfiles.R
\name{hashfiles}
\alias{hashfiles}
\title{Hash Files}
\usage{
hashfiles(
  x,
  skip.missing = FALSE,
  full.hash = FALSE,
  verbose = FALSE,
  skiptemp = TRUE
)
}
\arguments{
\item{x}{Input which specifies which files to hash. This can be a vector mix of paths and files.}

\item{skip.missing}{Skip missing files. Default is to throw an error if a file isn't found.}

\item{full.hash}{By default we just hash the file info (name, size, created/modified time). Set this to TRUE to read the file and hash the contents.}

\item{verbose}{Print helpful messages from code.}

\item{skiptemp}{Skip temporary MS Office files like "~$Simd Loss Eval 2018-06-30.xlsx"}
}
\value{
String representing hash of files.
}
\description{
Get a hash value representing a list of files. Useful for determining if files have changed in order to reset dependent caches.
}
\examples{
folder = system.file('extdata', package = 'easyr')
hashfiles(folder)
}
