% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.any.R
\name{headers_row}
\alias{headers_row}
\title{Identify headers row.}
\usage{
headers_row(
  x,
  headers_on_row = NA,
  first_column_name = NA,
  field_name_map = NA
)
}
\arguments{
\item{x}{Data frame to work with.}

\item{headers_on_row}{The specific row with headers on it.}

\item{first_column_name}{A known column(s) that can be used to find the header row. This is more flexible, but only used if headers_on_row is not available. If multiple are possible, use a vector argument here.}

\item{field_name_map}{field_name_map from read.any.}
}
\value{
List with headers_already_column_names (TRUE/FALSE); headers_on_row (1-indexed number of the to match standard R indexing).
}
\description{
Identify the row with headers in a data frame.
It should NOT be used directly (that's why it isn't exported), but will be called by function [read.any] as necessary, with the applicable defaults set by that function.
}
