% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_var.R
\name{histogram_var}
\alias{histogram_var}
\title{Create a histogram to visualize the distribution of a numeric variable.}
\usage{
histogram_var(data, var_name)
}
\arguments{
\item{data}{A data frame containing the variable to be plotted.}

\item{var_name}{The name of the variable to be visualized.}
}
\value{
A ggplot object with the histogram.
}
\description{
This function generates a histogram to visualize the distribution of a numeric variable
from a data frame.
}
\examples{
my_data <- data.frame(values = c(1, 2, 3, 4, 5))
histogram_var(my_data, "values")

}
