% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs.R
\name{ihs}
\alias{ihs}
\title{Inverse hyperbolic sine}
\usage{
ihs(x)
}
\arguments{
\item{x}{a numeric or complex vector.}
}
\value{
A numeric vector of transformed values.
}
\description{
Compute the inverse hyperbolic sine.
}
\examples{
x <- 1:1e4

plot(ihs(x), pch = 20, cex = 0.2, col = "red")
points(log(x), pch = 20, cex = 0.1, col = "black")

x <- rnorm(1e5, mean = 0, sd = 100) ^ 2
plot(density(x))

plot(density(ihs(x)))
lines(density(log1p(x)), col = "blue")
lines(density(log(x)), col = "red")
}
