% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_all_unique.R
\name{is_all_unique}
\alias{is_all_unique}
\title{Determine uniqueness of column}
\usage{
is_all_unique(x)
}
\arguments{
\item{x}{A vector or data frame column}
}
\value{
Boolean, whether or not the vector contains only unique values
}
\description{
Use to easily evaluate whether a data frame column or vector only contains unique values.
}
\author{
Kelli-Jean Chun, \email{kjchunz@gmail.com}
}
