% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{ischar}
\alias{ischar}
\title{Shorthand for is.character}
\usage{
ischar(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
logical indicator
}
\description{
Shorthand for is.character
}
\examples{
ischar( 'a character' )
ischar(1)
}
