% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{isdate}
\alias{isdate}
\title{Shorthand for lubridate::is.Date}
\usage{
isdate(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
logical indicator
}
\description{
Shorthand for lubridate::is.Date
}
\examples{
isdate( lubridate::mdy( '10/1/2014' ) )
isdate(1)
}
