% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{isfac}
\alias{isfac}
\title{Shorthand for is.factor}
\usage{
isfac(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
logical indicator
}
\description{
Shorthand for is.factor
}
\examples{
isfac( factor( c( 'a', 'b', 'c' ) ) )
isfac(1)
}
