% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{isnum}
\alias{isnum}
\title{Shorthand for is.numeric}
\usage{
isnum(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
logical indicator
}
\description{
Shorthand for is.numeric
}
\examples{
isnum(1)
isnum( factor( c( 'a', 'b', 'c' ) ) )
}
