% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_search.R
\name{jaccard_search}
\alias{jaccard_search}
\title{Jaccard Similarity Search}
\usage{
jaccard_search(search, context, level = 0.5, return_similarity = FALSE)
}
\arguments{
\item{search}{Single character/string to search for.}

\item{context}{Vector of characters to search within.}

\item{level}{Returned characters will be this similar or more similar. Higher values will return fewer/closer matches.}

\item{return_similarity}{Special option for diagnosing. TRUE will ignore [level] and return a named vector where the name is the context value and the value is the similarity.}
}
\value{
Characters that meet the similarity requirement.
}
\description{
Search for similar strings using Jaccard Similarity.
}
\examples{
jaccard_search('foobar', c('foo', 'bar', 'foobars'))
}
