% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_unique.R
\name{length_unique}
\alias{length_unique}
\title{Simple function to shorten length(unique(df$var))}
\usage{
length_unique(df_var)
}
\arguments{
\item{df_var}{What you want to know the length of unique values for: df$var}
}
\value{
Results of query
}
\description{
Simple function to shorten length(unique(df$var))
}
