% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_pad.R
\name{lz_pad}
\alias{lz_pad}
\title{Pad with zeroes on the left}
\usage{
lz_pad(string, width)
}
\arguments{
\item{string}{a vector.}

\item{width}{minimum number of characters after padding.}
}
\value{
A character vector.
}
\description{
This just calls \code{\link[stirngr]{str_pad}} to
pad each element of a character vector with zeros on the left side.
This is helpful when standardizing the length of numeric ID variables.
}
