% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff.R
\name{mdiff}
\alias{mdiff}
\title{Date Difference in Months}
\usage{
mdiff(x, y, do.date.convert = TRUE, do.numeric = TRUE)
}
\arguments{
\item{x}{Vector of starting dates or items that can be converted to dates by todate.}

\item{y}{Vector of ending dates or items that can be converted to dates by todate.}

\item{do.date.convert}{Convert to dates before running the difference. If you know your columns are already dates, setting to FALSE will make your code run faster.}

\item{do.numeric}{Convert the output to a number instead of a date difference object.}
}
\value{
Vector of differences.
}
\description{
Date Difference in Months
}
\examples{
mdiff( lubridate::mdy( '1/1/2018' ), lubridate::mdy( '3/4/2018' ) )
}
