% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left-right-mid.R
\name{mid}
\alias{mid}
\title{mid}
\usage{
mid(string, start, nchars)
}
\arguments{
\item{string}{String to process.}

\item{start}{Index (1-index) to start at.}

\item{nchars}{Number of characters to read in from start.}
}
\description{
Behaves like Excel's LEFT, RIGHT, and MID functions
Author: Bryce Chamberlain.
}
\examples{
mid( "leftmidright", 5, 3 )
}
