% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_corr.R
\name{myfunct_corr}
\alias{myfunct_corr}
\title{Title Get highly correlated variables with another given variable}
\usage{
myfunct_corr(dataset, field_of_interest, plot = FALSE,
  corr_type_plt = "pearson", plot_n = 10)
}
\arguments{
\item{dataset}{Dataframe}

\item{field_of_interest}{The variable you want to compare against.  Should be numerical}

\item{plot}{Binary.  Yes or No.  If yes returns a list of metrics and plot}

\item{corr_type_plt}{Options are 'pearson', 'spearman','kendall'.  Default is pearson.  Spell correctly}

\item{plot_n}{How many do you want to plot?}
}
\value{
Returns a list if plot is true.  Otherwise just returns the metrics.
}
\description{
Quickly get the highest correlated variables with another variable across the difference correlation methods
}
\examples{
corrs =myfunct_corr(mtcars, 'mpg', plot=TRUE)
}
