% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_crosstab.R
\name{myfunct_crosstab}
\alias{myfunct_crosstab}
\title{Title  Cross Tab (Table)}
\usage{
myfunct_crosstab(dataset, row_field, column_field,
  percent_denominator = "row", rounding_digits = 0)
}
\arguments{
\item{dataset}{Dataframe}

\item{row_field}{Ideally place the grouping variable here.}

\item{column_field}{Ideally place the target or outcome interested in comparinghere.}

\item{percent_denominator}{What is the demoninator for percents?  Row is default so place target in column}

\item{rounding_digits}{Default is 0 so whole percents are shown}
}
\value{
Returns the cross tab results...
}
\description{
Generate a cross table of the fields you are interested in.  Input of fields should be with names in quotes or as a result of a dplyr pipe similar to mtcars %>% select(vs).  dataset$field is not supported
}
\examples{
 myfunct_crosstab(mtcars, mtcars \%>\% select(cyl), mtcars \%>\% select(vs))
myfunct_crosstab(mtcars, 'cyl', 'vs')
}
