% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_data_audit.R
\name{myfunct_data_audit}
\alias{myfunct_data_audit}
\title{Title Data Audit}
\usage{
myfunct_data_audit(dataset, custom_metric = NULL,
  export_name = "data_audit.csv", run_nzv = TRUE, nzv_ratio = 50)
}
\arguments{
\item{dataset}{Dataframe.  Pass in dataframe %>% filter(your_col == <your group> if you just want to subset on a single group.)}

\item{custom_metric}{Enter with ()s and a dot for summarise_all.  For example, "sum(. > 100, na.rm = TRUE)" otherwise this will cause an error. Default is NULL for no custom metric}

\item{export_name}{file name to save metrics...}

\item{run_nzv}{run NZV function from caret? Default is true.}

\item{nzv_ratio}{what should the ratio be for the nzv function? caret default is 1:19 or >95%.  Default here is 50 or >98%.}
}
\value{
Returns a dataframe of metrics and saves CSV file as well...
}
\description{
Get an audit of your data with various metrics such as different quantiles, 0s, number of distincts, most common value, etc...
}
\examples{
 myfunct_data_audit(Sonar, custom_metric = "sum(. > 5, na.rm = TRUE)")
}
