% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_data_split.R
\name{myfunct_data_split}
\alias{myfunct_data_split}
\title{Title Split Data by Percentage}
\usage{
myfunct_data_split(dataset, train_pct, return_as_single_df = FALSE,
  seed = 2018, dataset_flag_name = c("train", "test"))
}
\arguments{
\item{dataset}{dataframe that you wish to divide}

\item{train_pct}{a number between 0 and 1.  This will be the percent of records from the dataset in the 'train' dataframe.}

\item{return_as_single_df}{Default is FALSE.  Return the dataset as a single dataframe with a column 'dataset_flag_name' indicating which record each belongs too}

\item{seed}{Default is 2018.  Ensures consistent splits}

\item{dataset_flag_name}{What do you want to call the dataset splits?  Default is train and test.}
}
\value{
Returns a dataframe with a new column indicating the records assignment or a list of dataframes if the return_as-single_df is false
}
\description{
Title Split Data by Percentage
}
\examples{

myfunct_data_split(iris, train_pct = .45)
myfunct_data_split(iris, train_pct = .45, return_as_single_df = TRUE)
}
