% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_eda_numerical.R
\name{myfunct_explore_numerical}
\alias{myfunct_explore_numerical}
\title{Title}
\usage{
myfunct_explore_numerical(dataset, select_variables = NULL,
  target_compare = NULL, specific_value_sum = NULL)
}
\arguments{
\item{dataset}{dataframe}

\item{select_variables}{specific variables only?}

\item{target_compare}{Target column if you want to produce pearson correlations}

\item{specific_value_sum}{Check how many values are less than greater than or equal to a particular value such as -1 or >100.   You must pass in the sign and value in quotes.  For example, ' < 0' or ' == 999'}
}
\value{
Returns a dataframe of metrics
}
\description{
Title
}
\examples{
myfunct_explore_numerical(mtcars)
}
