% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_gainslift.R
\name{myfunct_gainslift_funct}
\alias{myfunct_gainslift_funct}
\title{Title Gains, Lift, and KS Calculator}
\usage{
myfunct_gainslift_funct(df, binary_target, prediction)
}
\arguments{
\item{df}{Dataframe of prediction results}

\item{binary_target}{Should work for either numeric or factor.  Must be 0s and 1s though.}

\item{prediction}{The prediction value.   Higher values = higher chance of obtaining a 1}
}
\value{
Returns a dataframe with various metrics at each progression into the scorecard.
}
\description{
Outputs a dataframe that has the cumulative capture rate, gains, and KS metrics at each 1% increment of the score.   Also produces the lower threshold of the score.
}
