% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_get_date_values.R
\name{myfunct_get_date_values}
\alias{myfunct_get_date_values}
\title{Title Get Date Information}
\usage{
myfunct_get_date_values(dataset, column)
}
\arguments{
\item{dataset}{A dataframe}

\item{column}{the column that is in date format.}
}
\value{
Returns the original dataset with the date columns added
}
\description{
Quickly get a bunch of different date information
}
\examples{


}
