% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_getdummies.R
\name{myfunct_get_dummies}
\alias{myfunct_get_dummies}
\title{Title Get Dummy columns}
\usage{
myfunct_get_dummies(dataset, dummy_this_col,
  return_as_seperate_df = FALSE, drop_original_col = TRUE,
  convert_to_factor = TRUE)
}
\arguments{
\item{dataset}{Dataframe}

\item{dummy_this_col}{Column you wish to explode into new columns as dummy 0/1s}

\item{return_as_seperate_df}{Return the dummy columns on their own?  Default is FALSE so its added back to original dataframe}

\item{drop_original_col}{Drop originaly column?  Default is true}

\item{convert_to_factor}{Convert to factor?  Default is true}
}
\value{
Return a dataframe
}
\description{
Get dummy columns for a variable and append them back to your dataset.  Optionally drop the original column, convert new binary columns to factors, or return just binary columns seperately.  Keep new columns as integer if you want to calculate mean easily to see percentage of a given factor level.  New column names are given as  original_column.level
}
\examples{

myfunct_get_dummies(mtcars, 'cyl',return_as_seperate_df = FALSE, convert_to_factor = FALSE, drop_original_col = TRUE)
}
