% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_groupby_metric.R
\name{myfunct_groupby_metric}
\alias{myfunct_groupby_metric}
\title{Title Group-by Metrics}
\usage{
myfunct_groupby_metric(dataset, groupby_fields,
  metric = "mean(., na.rm=TRUE)", numeric_fields_only = TRUE,
  num_breaks = 5)
}
\arguments{
\item{dataset}{dataframe}

\item{groupby_fields}{one or a vector.  'vs' or c('vs','am')}

\item{metric}{Spell it out with understanding that it is being passed to summarize_all "mean(., na.rm=TRUE)"}

\item{numeric_fields_only}{By default filters out non-numeric.  Pass FALSE if you want to do something like n_distinct or sum(is.na(.))}

\item{num_breaks}{For Numerical Data, how many quantile bins?  Defaults to 5}
}
\value{
Returns a dataframe of all the grouping levels plus the entire dataset as whole
}
\description{
Quick group by metrics for a given group or groups.  Only Supports one for numerical values.
}
