% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_h2o_glm_coef.R
\name{myfunct_h2o_coefs}
\alias{myfunct_h2o_coefs}
\title{Title Get GLM coeffecients}
\usage{
myfunct_h2o_coefs(h2o_model)
}
\arguments{
\item{h2o_model}{Either the actual object or the string of the model name}
}
\value{
dataframe of coeffecients
}
\description{
Shortcut to getting h2o coeffecients from a glm model.  Retrieves all info around them that is available
}
