% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_h2osavemodel.R
\name{myfunct_h2osavemodel}
\alias{myfunct_h2osavemodel}
\title{Title Save H2o Model Information}
\usage{
myfunct_h2osavemodel(h2o_model, force = FALSE)
}
\arguments{
\item{h2o_model}{String or actual h2omodel}

\item{force}{Default is FALSE.  Controls whether to overwrite existing file...}
}
\value{
Nothing.  Saves models out working directory.  Force is FALSE,
}
\description{
Convience wrapper for saving h2o models.  Models can be saved whether they are passed as an actual model or if it is just the string of the name.  A folder will be created with the model information.
}
