% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_lookup_rename.R
\name{myfunct_lookup_rename}
\alias{myfunct_lookup_rename}
\title{Title Rename Columns by Lookup}
\usage{
myfunct_lookup_rename(df, df_column_lookup)
}
\arguments{
\item{df}{The datafraine}

\item{df_column_lookup}{a dataframe with a new and old column for the spelling of names}
}
\description{
Rename columns in a dataframe by a lookup table.
}
\details{
Lookup table must have the column names 'old' and 'new'
}
