% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_pairwise_correlations.R
\name{myfunct_pairwise_correlations}
\alias{myfunct_pairwise_correlations}
\title{Title Pairwise Correlation Creation}
\usage{
myfunct_pairwise_correlations(dataset, correlation_type = "pearson",
  reduce_data = TRUE)
}
\arguments{
\item{dataset}{Dataframe}

\item{correlation_type}{'spearman','kendall' or 'pearson'.  'pearson' is the default value.}

\item{reduce_data}{True by default.  Makes it so every combination can appear just once.  This might make it harder to find a given column by filtering since you won't know which column the variable-pair combination will appear in.    FALSE lets the variable exist in both var1 and var2 columns}
}
\value{
A 3 column dataframe sorted in descending absolute correlation order... 'var1','var2', and 'correlation'
}
\description{
Generate pairwise correlation combinations.  This makes it easier to examine in some circumstances rather than
}
\examples{

myfunct_pairwise_correlations(dataset='mtars',correlation_type='spearman', reduce_data=TRUE)
}
