% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_range01.R
\name{myfunct_range01}
\alias{myfunct_range01}
\title{Title Transform a value from range 0 to 1}
\usage{
myfunct_range01(var)
}
\arguments{
\item{var}{the variable that you want to transform.}
}
\value{
Returns the transformed values
}
\description{
Easily rescale a value from 0 to 1.
}
\examples{

myfunct_range01(mtcars$hp)
myfunct_range01(mtcars \%>\% select(hp))

}
