% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunct_rformula.R
\name{myfunct_rformula}
\alias{myfunct_rformula}
\title{Title  R Formula Creation}
\usage{
myfunct_rformula(x, y)
}
\arguments{
\item{x}{The x variables in your model.}

\item{y}{The y variable in your model.}
}
\value{
Returns a formula that you can plug into R models such as lm(returned_formula,  data=mtcars)
}
\description{
Easily create a formula for R models.  This can help save time when there are many different variables as you can pass them in as a vector.
}
\examples{

}
