% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr_h2o_filter.R, R/h2ofilter.R
\name{myh2o.filter}
\alias{myh2o.filter}
\title{H2o Subset Rows}
\usage{
myh2o.filter(h2o_df, column_to_filter, sign, value)

myh2o.filter(h2o_df, column_to_filter, sign, value)
}
\arguments{
\item{h2o_df}{h2odataframe}

\item{column_to_filter}{name of column to filter using}

\item{sign}{should be >=, >, <, <=, ==, !=.   In future will add null/not null}

\item{value}{what value is being compared.}

\item{h2o_df}{h2odataframe}

\item{column_to_filter}{name of column to filter using}

\item{sign}{should be >=, >, <, <=, ==, !=.   In future will add null/not null}

\item{value}{what value is being compared.}
}
\value{
H2o dataframe after filtering is applied

H2o dataframe after filtering is applied
}
\description{
Convience wrapper for avoiding syntax issues when filtering using H2O

Convience wrapper for avoiding syntax issues when filtering using H2O
}
