% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot_cum_dens.R
\name{myplot_cum_density}
\alias{myplot_cum_density}
\title{Title Plot Cumulative Amount of a Variable}
\usage{
myplot_cum_density(dataset, numeric_field, grouping_field = NULL,
  title = NULL)
}
\arguments{
\item{dataset}{Dataframe}

\item{numeric_field}{Variable you want to plot}

\item{grouping_field}{The grouping variable. NULL by default}

\item{title}{Title of plot.}
}
\value{
Returns a plot
}
\description{
Plot the cumulative percentage against a variable's value.
}
\examples{

myplot_cum_density(Sonar, 'V11')
myplot_cum_density(Sonar, numeric_field='V11', grouping_field='Class')
}
