% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_na.R
\name{switch_na}
\alias{switch_na}
\alias{zero_na}
\alias{na_zero}
\alias{null_na}
\title{Switch out missing values}
\usage{
zero_na(x)

na_zero(x)

null_na(x, unlist = FALSE, ...)
}
\arguments{
\item{x}{a vector.}

\item{unlist}{logical, whether to unlist the result.}

\item{...}{further arguments to pass to \code{unlist()}.}
}
\value{
For \code{zero_na}, \code{0} values replaced with \code{NA}.

For \code{na_zero}, \code{NA} values replaced with \code{0}.

For \code{null_na}, \code{NULL} values replaced with \code{NA}.
}
\description{
Switch out missing values
}
\examples{
x <- sample(c(0:5, NA), 10, replace = TRUE)
rbind(x,
      zero_na = zero_na(x),
      na_zero = na_zero(x)
      )

y <- list(1, 0, NULL, 1, 0)
null_na(y)
}
