% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nastrings}
\alias{nastrings}
\title{NA Strings}
\format{
A vector of values.
}
\usage{
nastrings
}
\description{
A list of strings to consider NA. Includes blank string, "NA", excel errors, etc. Used throughout easyr for checking NA.
}
\keyword{datasets}
