% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{null}
\alias{null}
\title{Shorthand for is.null}
\usage{
null(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
logical indicator
}
\description{
Shorthand for is.null
}
\examples{
null( NULL )
null(1)
}
