% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\alias{detect_outliers}
\alias{rm_outliers}
\title{Detect and remove outliers}
\usage{
detect_outliers(x, probs, na.rm = FALSE, incbounds = TRUE)

rm_outliers(df, ..., probs = c(0.025, 0.975), na.rm = FALSE,
  incbounds = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{probs}{for \code{detect_outliers()}, a length-2 numeric vector of probabilities. For \code{rm_outliers()}, the same or a list of such vectors. See \code{\link[base]{quantile}}.}

\item{na.rm}{logical, whether to remove \code{NA} values before computing quantiles.}

\item{incbounds}{logical, whether boundary values should be interpreted as inclusive. See \code{\link[data.table]{between}}.}

\item{df}{a \code{data.frame}.}

\item{...}{quoted column names in which to search for outliers, or a vector of column names.}
}
\description{
Detect and remove outliers, as defined by quantile probabilities.
}
\examples{
df <- data.frame(num1 = 1:1e3, num2 = sample(1:1e3))

df$num1[detect_outliers(df$num1, probs = c(0.025, 0.975))]

identical(rm_outliers(df, c("num1", "num2")),
          rm_outliers(df, "num1", "num2")
          )
}
\seealso{
\code{\link[base]{quantile}} and \code{\link[data.table]{between}}.
}
