% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linear_regression.R
\name{plot_linear_regression}
\alias{plot_linear_regression}
\title{Create a scatter plot with a linear regression line.}
\usage{
plot_linear_regression(data, x_var, y_var)
}
\arguments{
\item{data}{A data frame containing the variables to be plotted.}

\item{x_var}{The name of the variable for the x-axis.}

\item{y_var}{The name of the variable for the y-axis.}
}
\value{
A ggplot object with the scatter plot and the linear regression line.
}
\description{
This function generates a scatter plot between two numeric variables and fits
a linear regression line to the data.
}
\examples{
my_data <- data.frame(x = c(1, 2, 3, 4, 5), y = c(2, 4, 5, 4, 6))
plot_linear_regression(my_data, "x", "y")

}
