% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot percent missing for all variables in DF}
\usage{
plot_missing(df)
}
\arguments{
\item{df}{Data frame to be used}
}
\value{
Primary: A plot of missing percentage for every variable in DF

Secondary: You can extract the missing % data by assigning function
to an object and pulling out the data (e.g. the_object$data)
}
\description{
This function will plot the percentage of missing observations for
every variable in a data frame.
}
