% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_na.R
\name{prop_na}
\alias{prop_na}
\alias{count_na}
\title{Calculate the proportion of missing values}
\usage{
prop_na(x)

count_na(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A scalar. For \code{prop_na} the proportion of missing  values. For \code{count_na} a count of missing values
}
\description{

}
