% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.any.R
\name{rany_fixColNames}
\alias{rany_fixColNames}
\title{Fix column names.}
\usage{
rany_fixColNames(col_names, fix.dup.column.names, nastrings)
}
\arguments{
\item{col_names}{Vector/value of colum names/name.}

\item{fix.dup.column.names}{Adds 'DUPLICATE #' to duplicated column names to avoid errors with duplicate names.}

\item{nastrings}{Characters/strings to read as NA.}
}
\value{
Fixed names.
}
\description{
Code to fix column names, since this has to be done up to twice will reading in files.
It should NOT be used directly (that's why it isn't exported), but will be called by function [read.any] as necessary, with the applicable defaults set by that function.
}
