% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.txt.R
\name{read.txt}
\alias{read.txt}
\title{Read File as Text}
\usage{
read.txt(filename, folder = NA)
}
\arguments{
\item{filename}{File path and name for the file to be read in.}

\item{folder}{Folder path to look for the file in.}
}
\value{
Character variable containing the text in the file.
}
\description{
Read File as Text
}
\examples{
# read.txt( 'tests/testthat/test-files/some-text.txt' )
}
