% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{map}
\alias{map_at}
\alias{str_extract}
\alias{tribble}
\alias{skim}
\alias{tabyl}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{tidy}}}

  \item{janitor}{\code{\link[janitor]{tabyl}}}

  \item{purrr}{\code{\link[purrr]{map}}, \code{\link[purrr]{map_at}}}

  \item{skimr}{\code{\link[skimr]{skim}}}

  \item{stringr}{\code{\link[stringr]{str_extract}}}

  \item{tibble}{\code{\link[tibble]{tribble}}}
}}

