% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_legend.R
\name{remove_legend}
\alias{remove_legend}
\title{Remove the legend}
\usage{
remove_legend(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{ggplot that has a legend that you want to remove}
}
\value{
The ggplot's legend
}
\description{
Returns a ggplot object, with the legend removed.
}
\examples{
library(ggplot2)
library(gridExtra)
library(grid)
cars = mtcars
cars$cyl = factor(cars$cyl)
p1 = ggplot(cars, aes(x=wt, y=mpg, group=cyl, colour=cyl)) +
  geom_point() +
  theme(legend.position="bottom")
p2 = ggplot(cars, aes(x=wt, y=hp, group=cyl, colour=cyl)) +
  geom_point()
legend = get_legend(p1)
grid.arrange(arrangeGrob(remove_legend(p1),
                         remove_legend(p2),
                         ncol=2),
             legend,
             heights=c(10,1))

}
\author{
Kelli-Jean Chun, \email{kjchunz@gmail.com}
}
