% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_setdiff.R
\name{rev_setdiff}
\alias{rev_setdiff}
\title{Reverse set difference}
\usage{
rev_setdiff(x, y)
}
\arguments{
\item{x, y}{vectors.}
}
\value{
The output of \code{setdiff(y, x)}.
}
\description{
Same as \code{\link[base]{setdiff}} with argument order reversed.
}
\details{
This is a convenience function useful when \code{x} or \code{y} is long/cumbersome to type.
}
\examples{
x <- LETTERS[1:10]
y <- LETTERS[5:14]

setdiff(x, y)
rev_setdiff(y, x)
rev_setdiff(x, y)

}
\seealso{
\code{\link[base]{setdiff}}.
}
