% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{right}
\usage{
right(string, char)
}
\arguments{
\item{string}{String to process.}

\item{char}{Number of characters.}
}
\description{
Behaves like Excel's left() and right() functions
Author: Dave. Tech review: Bryce Chamberlain.
}
\examples{
right("leftright",5)
}
