% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_tmp.R
\name{rm_tmp}
\alias{rm_tmp}
\title{Remove temporary files}
\usage{
rm_tmp(pattern = NULL)
}
\description{
Remove files in the \R session's temporary directory.
This function is useful in conjunction with packages like \code{raster} and \code{aws.s3}, which frequently write files to \code{tempdir()}.
}
\examples{
tmp_csv <- tempfile(fileext = rep(c(".csv", ""), 10))
lapply(tmp_csv, function(path) write.csv("", path))

list.files(tempdir())

rm_tmp(pattern = "\\\\.csv")
list.files(tempdir())

rm_tmp()
list.files(tempdir())

}
