% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.any.R
\name{rx}
\alias{rx}
\title{Read Excel}
\usage{
rx(filename, sheet, first_column_name, nrows, verbose)
}
\arguments{
\item{filename}{File path and name for the file to be read in.}

\item{sheet}{The sheet to read in.}

\item{first_column_name}{Pass a column name to help the function find the header row.}

\item{nrows}{Number of rows to read in.}

\item{verbose}{Print helpful messages via cat().}
}
\value{
Data object
}
\description{
This gets a bit complex since many errors can occur when reading in excel files. We've done our best to handle common ones.
Requires packages: openxlsx, readxl, XML (these are required by easyr).
It should NOT be used directly (that's why it isn't exported), but will be called by function [read.any] as necessary, with the applicable defaults set by that function.
}
