% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list.R
\name{save_list}
\alias{save_list}
\title{Write list elements to disk as separate objects}
\usage{
save_list(x, dir, ext = c(".rds", ".csv"), names = NULL, cores = 1,
  ...)
}
\arguments{
\item{x}{a list.}

\item{dir}{a directory in which to save the list elements.}

\item{ext}{a file extension.}

\item{names}{a vector of file names.}

\item{cores}{number of cores to use in \code{\link{mcparallel}}. The default (1) runs in series.}

\item{...}{further arguments to pass to \code{\link{saveRDS}} or \code{\link[data.table]{fwrite}}.}
}
\description{
Write list elements to disk as separate objects
}
\examples{

tmp_dir <- tempdir()
df <- data.frame(char = LETTERS[1:10], num = 1:10)
ldf <- replicate(3, df, simplify = FALSE)

save_list(ldf, dir = tmp_dir, ext = ".csv")

list.files(tmp_dir)
unlink(tmp_dir, recursive = TRUE)
}
\seealso{
\code{\link{saveRDS}} and \code{\link[data.table]{fwrite}}.
}
