% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_case.R
\name{set_case}
\alias{set_case}
\alias{set_upper}
\alias{set_lower}
\title{Set all character columns of a data.table to a type case}
\usage{
set_upper(DT)

set_lower(DT)
}
\arguments{
\item{DT}{a \code{data.table}, or something coercible to one. See \code{\link[data.table]{setDT}}.}
}
\description{
Set all character columns of a data.table to a type case
}
\examples{

df <- data.frame(a = letters, b = LETTERS,
                 stringsAsFactors = FALSE)

set_lower(df)
set_upper(df)

}
\seealso{
\code{\link[base]{tolower}} and \code{\link[base]{toupper}}.
}
