% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cols.R
\name{set_cols}
\alias{set_cols}
\title{Internally looped \code{set()}}
\usage{
set_cols(x, FUN, i = NULL, j = NULL)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{FUN}{a function to be applied to all \code{x[i, j]}.}

\item{i}{integer indices indicating rows to update.}

\item{j}{column names or integer indices indicating columns to update.
The default updates all columns.}
}
\description{
A convenience function implementing \code{\link[data.table]{set}}
looped over columns of a \code{data.frame}.
Modification is fast (by reference) but cannot include grouping operations.
}
\seealso{
\code{\link[data.table]{:=}}.
}
