% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_max_col.R
\name{set_max_col}
\alias{set_max_col}
\title{Re-establishes the lengths of variables in a sample data frame to be the max(length) of
the full data frame
NOTE: Currently only sets sizes for character variables}
\usage{
set_max_col(full_df, samp_df)
}
\arguments{
\item{full_df}{the original, non-sampled, data frame}

\item{samp_df}{name of data frame to change lengths}
}
\value{
samp_df with changed lengths
}
\description{
Re-establishes the lengths of variables in a sample data frame to be the max(length) of
the full data frame
NOTE: Currently only sets sizes for character variables
}
