% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_dir.R
\name{setwd_cur}
\alias{setwd_cur}
\title{Set the wd to the current directory of the file}
\usage{
setwd_cur()
}
\description{
Set the wd to the current directory of the file
}
\examples{
setwd_cur()
}
\seealso{
\code{\link{cur_dir}}, \code{\link{cur_dir_source}}
}
\author{
Kelli-Jean Chun, \email{kjchunz@gmail.com}
}
