% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_text.R
\name{similar_text}
\alias{similar_text}
\title{Text Similarity Search}
\usage{
similar_text(
  search,
  context,
  algo = "jaccard",
  level = 0.5,
  return_similarity = FALSE
)
}
\arguments{
\item{search}{Single character/string to search for.}

\item{context}{Vector of characters to search within.}

\item{algo}{Algorithm to use when determining similarity. Currenly, only Jaccard Similarity is implemented.}

\item{level}{Returned characters will be this similar or more similar. Higher values will return fewer/closer matches.}

\item{return_similarity}{Special option for diagnosing. TRUE will ignore [level] and return a named vector where the name is the context value and the value is the similarity.}
}
\value{
Characters that meet the similarity requirement.
}
\description{
Search for similar strings using in a vector.
}
\examples{
similar_text('foobar', c('foo', 'bar', 'foobars'))
similar_text('foobar', c('foo', 'bar', 'foobars'), return_similarity = TRUE)
}
