% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_lower.R
\name{snake_lower}
\alias{snake_lower}
\title{Renames variables using snake(_) and lower syntax}
\usage{
snake_lower(df)
}
\arguments{
\item{data}{Data frame to be renamed}
}
\value{
Date frame returned with renamed variables
}
\description{
Renames variables using snake(_) and lower syntax
}
