% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl.R
\name{spl}
\alias{spl}
\title{Sample}
\usage{
spl(x, n = 10, warn = TRUE, replace = FALSE, ...)
}
\arguments{
\item{x}{Data to sample from.}

\item{n}{Number or percentage of rows/values to return. If less than 1 it will be interpreted as a percentage.}

\item{warn}{Warn if sampling more than the size of the data.}

\item{replace}{Whether or not to sample with replacement.}

\item{...}{Other parameters passed to sample()}
}
\value{
Sample dataframe/vector.
}
\description{
Extracts a uniform random sample from a dataset or vector. Provides a simpler API than base R.
Author: Bryce Chamberlain. Tech Review: Maria Gonzalez.
}
\examples{
spl( c(1:100) )
spl( c(1:100), n = 50 )
spl( iris )
}
