% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert_big.R
\name{sql_insert_big}
\alias{sql_insert_big}
\title{Uses the `sql_insert` function from `mmkit`, but cuts up inserting into batch sizes
when table is too large for efficient insert.}
\usage{
sql_insert_big(
  df_name,
  table_name,
  batch_size,
  delete_phantom = FALSE,
  phantom_null_field = "insert_date"
)
}
\arguments{
\item{df_name}{name of data frame to insert}

\item{table_name}{Redshift table name (with schema if required)}

\item{batch_size}{Batch size for inserts (size of subset chunks)}
}
\value{
Results of query
}
\description{
Uses the `sql_insert` function from `mmkit`, but cuts up inserting into batch sizes
when table is too large for efficient insert.
}
