% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strx.R
\name{strx}
\alias{strx}
\title{Structure with Like}
\usage{
strx(df, char, ignore.case = T)
}
\arguments{
\item{df}{Object with names you'd like to search.}

\item{char}{Regex (character value) to match.}

\item{ignore.case}{(Optional) Ignore case when matching.}
}
\description{
Runs str function but only for names matching a character value (regex).
Author: Scott Sobel. Tech Review: Bryce Chamberlain.
}
\examples{
strx(iris,'length')
}
