% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_url.R, R/table_na.R
\name{table_na}
\alias{table_na}
\title{Simple function to shorten URL's to their simple base
Example: https://www.facebook.com/ --> facebook.com}
\usage{
table_na(df_var)

table_na(df_var)
}
\arguments{
\item{df_var}{What you want to know the # of NA & non-NA for: df$var}

\item{long_url}{The original URL}
}
\value{
base_url

Results of query
}
\description{
Simple function to shorten URL's to their simple base
Example: https://www.facebook.com/ --> facebook.com

Simple function to shorten table(is.na(df$var))
}
