% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{tochar}
\alias{tochar}
\title{Shorthand for as.character}
\usage{
tochar(x)
}
\arguments{
\item{x}{Value to check.}
}
\value{
as.character result
}
\description{
Shorthand for as.character
}
\examples{
tochar(NA)
tochar(1)
}
