% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usepkg.R
\name{usepkg}
\alias{usepkg}
\title{Use Package}
\usage{
usepkg(packages, noCache = FALSE, repos = "http://cran.us.r-project.org")
}
\arguments{
\item{packages}{Character or character vector with names of the packages you want to use.}

\item{noCache}{When checking packages, you can choose to ignore the cached list, which will increase accuracy but decrease speed.}

\item{repos}{choose the URL to install from.}
}
\description{
Installs a package if it needs to be installed, and calls require to load the package.
Author: Scott Sobel. Tech Review: Bryce Chamberlain.
}
\examples{

# packages shouldn't be installed during tests or examples according to CRAN. 
# therefore, examples cannot be provided because CRAN now runs donttest examples.

usepkg('geodist', FALSE, 'http://cran.us.r-project.org')

}
