% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff.R
\name{ydiff}
\alias{ydiff}
\title{Date Difference in Years}
\usage{
ydiff(x, y, do.date.convert = TRUE, do.numeric = TRUE)
}
\arguments{
\item{x}{Vector of starting dates or items that can be converted to dates by todate.}

\item{y}{Vector of ending dates or items that can be converted to dates by todate.}

\item{do.date.convert}{Convert to dates before running the difference. If you know your columns are already dates, setting to FALSE will make your code run faster.}

\item{do.numeric}{Convert the output to a number instead of a date difference object.}
}
\value{
Vector of differences.
}
\description{
Date Difference in Years
}
\examples{
ydiff( lubridate::mdy( '1/1/2018' ), lubridate::mdy( '3/4/2018' ) )
}
