% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{begin}
\alias{begin}
\title{Begin}
\usage{
begin(wd = NULL, load = c("magrittr", "dplyr"), scipen = FALSE, verbose = TRUE)
}
\arguments{
\item{wd}{Path to set as working directory. If blank, the location of the current file open in RStudio will be used if available. If FALSE, the working directory will not be changed.}

\item{load}{Packages to load. If not available, they'll be installed.}

\item{scipen}{Do scientific notation in output?}

\item{verbose}{Print information about what the function is doing?}
}
\description{
Perform common operations before running a script. Includes clearing environment variables, disabling scientific notation, loading common packages, and setting the working directory to the location of the current file.
}
\examples{

begin()
}
