% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_cols.R
\name{comp_cols}
\alias{comp_cols}
\title{Compare similarly named columns}
\usage{
comp_cols(x, id_cols = NULL, pattern = "^i\\\\.", col_list = NULL,
  mismatch_only = FALSE, match_only = FALSE, rm_na = FALSE)
}
\arguments{
\item{x}{a data.frame.}

\item{id_cols}{any ID columns to retain in each data.frame output.}

\item{pattern}{a regular expression pattern to match the distinguishing prefix/suffix.}

\item{col_list}{a list of \code{character(2)} column-name pairs, Non-\code{NULL} values here take precedence over \code{pattern}.}

\item{mismatch_only}{return only rows with different values in column pairs.}

\item{match_only}{return only rows with the same values in column pairs.}

\item{rm_na}{remove rows with \code{NA} in each set of comparison columns.}
}
\value{
A list of data.frame objects.
}
\description{
Extract columns with similar names and output a list of separate data.frames for easy comparison.
The presence/absence of a prefix/suffix identifies similar colums,
and the default corresponds to default-data.table joining (e.g., \code{i.count}, \code{count}).
You can provide a list of column names for extraction instead.
}
\examples{
x <- data.frame(char = LETTERS[1:10],
                i.char = sample(LETTERS[1:10]),
                num = 1:10,
                i.num = sample(1:10)
                )
comp_cols(x)

}
