% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_cols.R
\name{comp_cols}
\alias{comp_cols}
\title{Compare similarly named columns}
\usage{
comp_cols(x, id_cols = NULL, pattern = "^i\\\\.", col_list = NULL,
  restrict = FALSE, rm_na = FALSE)
}
\arguments{
\item{x}{a data.frame.}

\item{id_cols}{character vector of any ID columns to retain in each data.frame output.}

\item{pattern}{a regular expression pattern to match the distinguishing prefix/suffix.}

\item{col_list}{a list of \code{character(2)} column-name pairs, Non-\code{NULL} values here take precedence over \code{pattern}.}

\item{restrict}{one of \code{"match"}, \code{"mismatch"}, or \code{FALSE}, specifying whether to restrict the output to matches only, mismatches only, or not at all (the default).}

\item{rm_allna}{logical indicating whether, in each set of comparison columns, all-\code{NA} rows should be removed.}
}
\value{
A list of data.frame objects.
}
\description{
Extract columns with similar names and output a list of separate data.frames for easy comparison.
The presence/absence of a prefix/suffix identifies similar colums,
and the default corresponds to default-data.table joining (e.g., \code{i.count}, \code{count}).
You can provide a list of column names for extraction instead.
}
\examples{
set.seed(575)

x <- data.frame(id = paste0("row_", 1:10),
                char = LETTERS[1:10],
                i.char = sample(LETTERS[1:10]),
                num = 1:10,
                i.num = sample(1:10)
                )

comp_cols(x)
comp_cols(x, id_cols = "id", restrict = "mismatch")

x[2, "char"] <- NA
x[2, "num"] <- NA
x[3, c("char", "i.char")] <- NA
x[5, c("num", "i.num")] <- NA

comp_cols(x)
comp_cols(x, id_cols = "id", rm_na = TRUE)

}
