% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Link data.tables by fuzzy string matching}
\usage{
fuzzy_match(a, b, acol, bcol, blocks = NULL, method = "jw", ...)
}
\arguments{
\item{a}{a source data.table}

\item{b}{a target data.table}

\item{acol}{column name in \code{a} to use for matching}

\item{bcol}{column name in \code{b} to use for matching}

\item{blocks}{an optional character vector of column names referring to any variables to be used for \emph{exact} blocking.}

\item{method}{method for \code{\link[stringdist]{stringdistmatrix}}}
}
\value{
a data.table containing any blocking columns,
the source column,
the closest match in the target column,
and the string distance for that match.
}
\description{
Finds the closest string match between two data.tables.
The default method computes Jaro-Winkler string distances using the \code{stringdist} package.
In cases with multiple closest matches, only the first match is reported.
}
\examples{
library(data.table)

set.seed(575)
DTA <- data.table(block_a1 = sample(LETTERS[1:4], 20, TRUE),
                  block_a2 = sample(LETTERS[1:4], 20, TRUE),
                  fruit   = sample(stringr::fruit[1:12], 20, TRUE))

DTB <- data.table(block_b1 = sample(LETTERS[1:4], 20, TRUE),
                  block_b2 = sample(LETTERS[1:4], 20, TRUE),
                  fruit   = sample(stringr::fruit[1:12], 20, TRUE))

blocks <-  c(block_a1 = "block_b1", block_a2 = "block_b2")

fuzzy_match(DTA, DTB, "fruit", "fruit")
fuzzy_match(DTA, DTB, "fruit", "fruit", blocks = blocks)
}
\seealso{
\link[stringdist]{stringdistmatrix}
}
