% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left-right-mid.R
\name{left}
\alias{left}
\title{left}
\usage{
left(string, char)
}
\arguments{
\item{string}{String to process.}

\item{char}{Number of characters.}
}
\description{
Behaves like Excel's LEFT, RIGHT, and MID functions
Author: Dave. Tech review: Bryce Chamberlain.
}
\examples{
left( "leftmidright", 4 )
}
