% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{sch}
\alias{sch}
\title{Search a Data Frame.}
\usage{
sch(x, pattern, ignore.case = FALSE, fixed = FALSE)
}
\arguments{
\item{x}{Data to search.}

\item{pattern}{Regex patter to search. Most normal search terms will work fine, too.}

\item{ignore.case}{Ignore case in search (uses grepl).}

\item{fixed}{Passed to grepl to match string as-is instead of using regex. See ?grepl.}
}
\value{
Matching rows.
}
\description{
Searches all columns for a term and returns all rows with at least one match.
Author: Bryce Chamberlain.
}
\examples{
sch( iris, 'setosa' )
}
