% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_na.R
\name{table_na}
\alias{table_na}
\title{Simple function to shorten table(is.na(df$var))}
\usage{
table_na(df_var)
}
\arguments{
\item{df_var}{What you want to know the # of NA & non-NA for: df$var}
}
\value{
Results of query
}
\description{
Simple function to shorten table(is.na(df$var))
}
