% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usepkg.R
\name{usepkg}
\alias{usepkg}
\title{Use Package}
\usage{
usepkg(packages, noCache = FALSE)
}
\arguments{
\item{packages}{Character or character vector with names of the packages you want to use.}

\item{noCache}{When checking packages, you can choose to ignore the cached list, which will increase accuracy but decrease speed.}
}
\description{
Installs a package if it needs to be installed, and calls require to load the package.
Author: Scott Sobel. Tech Review: Bryce Chamberlain.
}
\examples{
usepkg(c('dplyr','lubridate'))
usepkg('lubridate')
}
